% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentGenomeBuild.R
\name{currentGenomeBuild}
\alias{currentGenomeBuild}
\alias{currentEnsemblGenomeBuild}
\alias{currentGencodeGenomeBuild}
\alias{currentRefseqGenomeBuild}
\alias{currentUcscGenomeBuild}
\title{Current genome build}
\usage{
currentEnsemblGenomeBuild(organism)

currentGencodeGenomeBuild(organism)

currentRefseqGenomeBuild(organism, taxonomicGroup = NULL)

currentUcscGenomeBuild(organism)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{taxonomicGroup}{\code{character(1)}.
\emph{Only applies to RefSeq}.
FTP server taxonomic group subdirectory path (e.g. "vertebrate_mammalian").
Defining this manually avoids having to query the FTP server.}
}
\value{
\code{character(1)}.
Genome assembly build version.
}
\description{
Fetch the current genome build (assembly) version from online resources.
}
\note{
Updated 2023-10-04.
}
\examples{
## Ensembl.
try({
    currentEnsemblGenomeBuild("Homo sapiens")
    currentEnsemblGenomeBuild("Mus musculus")
})

## GENCODE.
try({
    currentGencodeGenomeBuild("Homo sapiens")
    currentGencodeGenomeBuild("Mus musculus")
})

## RefSeq.
try({
    currentRefseqGenomeBuild(
        organism = "Homo sapiens",
        taxonomicGroup = "vertebrate_mammalian"
    )
    currentRefseqGenomeBuild(
        organism = "Mus musculus",
        taxonomicGroup = "vertebrate_mammalian"
    )
})

## UCSC.
try({
    currentUcscGenomeBuild("Homo sapiens")
    currentUcscGenomeBuild("Mus musculus")
})
}
\seealso{
\itemize{
\item \href{https://rest.ensembl.org}{Ensembl REST API}.
\item \href{https://genome.ucsc.edu/goldenPath/help/api.html}{UCSC REST API}.
\item \href{https://ftp.ncbi.nlm.nih.gov/genomes/refseq/}{RefSeq genomes}.
}
}
