% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentGenomeVersion.R
\name{currentGenomeVersion}
\alias{currentGenomeVersion}
\alias{currentEnsemblVersion}
\alias{currentGencodeVersion}
\alias{currentRefSeqVersion}
\alias{currentFlyBaseVersion}
\alias{currentWormBaseVersion}
\title{Current genome version}
\usage{
currentEnsemblVersion()

currentGencodeVersion(organism = c("Homo sapiens", "Mus musculus"))

currentRefSeqVersion()

currentFlyBaseVersion(dmel = FALSE)

currentWormBaseVersion()
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{dmel}{\code{logical(1)}.
Return \emph{Drosophila melanogaster} genome release version.}
}
\value{
\code{character(1)} or \code{integer(1)} when possible.
Release version.
}
\description{
Obtain the latest release version from various genome annotation sources.
}
\note{
Updated 2023-09-15.
}
\examples{
## Ensembl.
x <- try(
    expr = {
        currentEnsemblVersion()
    },
    silent = TRUE
)
print(x)
## GENCODE.
x <- try(
    expr = {
        currentGencodeVersion(organism = "Homo sapiens")
    },
    silent = TRUE
)
print(x)
## RefSeq.
x <- try(
    expr = {
        currentRefSeqVersion()
    },
    silent = TRUE
)
print(x)
## WormBase.
x <- try(
    expr = {
        currentWormBaseVersion()
    },
    silent = TRUE
)
print(x)
}
\seealso{
Refer to the koopa package for Bash shell variants.
}
