% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRefseqGenome.R
\name{downloadRefseqGenome}
\alias{downloadRefseqGenome}
\title{Download RefSeq reference genome}
\usage{
downloadRefseqGenome(
  organism,
  taxonomicGroup = NULL,
  genomeBuild = NULL,
  outputDir = getwd(),
  cache = FALSE,
  analysisSet = FALSE
)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{taxonomicGroup}{\code{character(1)}.
\emph{Only applies to RefSeq}.
FTP server taxonomic group subdirectory path (e.g. "vertebrate_mammalian").
Defining this manually avoids having to query the FTP server.}

\item{genomeBuild}{\code{character(1)}.
RefSeq genome build assembly name (e.g. \code{"GCF_000001405.39_GRCh38.p12"}).
If set \code{NULL}, defauls to the most recent build available.}

\item{outputDir}{\code{character(1)}.
Output directory path.}

\item{cache}{\code{logical(1)}.
Cache URLs locally, using BiocFileCache internally.}

\item{analysisSet}{\code{logical(1)}.
Download the NCBI "no-alt analysis set" FASTA alongside the standard
genome. This ALT-contig-free FASTA is recommended for aligners such as
STAR that do not natively handle ALT contigs. Currently supported for
GRCh38 and GRCm39. Defaults to \code{FALSE}.}
}
\value{
Invisible \code{list}.
}
\description{
Download RefSeq reference genome
}
\note{
Updated 2023-04-14.
}
\section{Stable release}{

The latest assembly defined under the "release/" subdirectory is not
considered "stable" by the RefSeq team. It is considered good practice to use
a genome build one version behind as a stable release
(e.g. "GCF_000001405.38_GRCh38.p12" instead of current
"GCF_000001405.39_GRCh38.p13" build).
}

\examples{
## This example is bandwidth intensive.
## > downloadRefseqGenome(
## >     organism = "Homo sapiens",
## >     taxonomicGroup = "vertebrate_mammalian",
## >     genomeBuild = "GCF_000001405.39_GRCh38.p12",
## > )
}
\seealso{
\itemize{
\item \href{https://www.ncbi.nlm.nih.gov/projects/genome/guide/human/}{Human Genome Resources at NCBI}
\item \href{https://ftp.ncbi.nlm.nih.gov/genomes/refseq/}{RefSeq Genomes FTP server}
\item \href{https://www.ncbi.nlm.nih.gov/genome/doc/ftpfaq/}{Genomes Download (FTP) FAQ}
\item \href{https://www.ncbi.nlm.nih.gov/genbank/genomes_gff/}{GenBank annotating genomes with GFF3 or GTF files}
\item \href{https://www.ncbi.nlm.nih.gov/genbank/collab/db_xref/}{db_xrefs}
\item \href{https://steinbaugh.com/posts/bcbio-hg38.html}{bcbio hg38 reference genome}
\item \href{https://lh3.github.io/2017/11/13/which-human-reference-genome-to-use}{Heng Li: Which human reference genome to use?}
\item \href{https://ftp.ncbi.nlm.nih.gov/genomes/all/GCA/000/001/405/GCA_000001405.15_GRCh38/seqs_for_alignment_pipelines.ucsc_ids/}{GRCh38 assembly for alignment pipelines}
\item \href{https://hgdownload.soe.ucsc.edu/goldenPath/hg38/bigZips/}{UCSC hg38 bigZips}
}
}
