% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Ensembl2Entrez-methods.R
\name{Ensembl2Entrez}
\alias{Ensembl2Entrez}
\alias{Ensembl2Entrez,GenomicRanges-method}
\alias{Ensembl2Entrez,character-method}
\title{Ensembl-to-Entrez gene identifier mappings}
\usage{
Ensembl2Entrez(object, ...)

\S4method{Ensembl2Entrez}{GenomicRanges}(object, format = c("1:1", "long"))

\S4method{Ensembl2Entrez}{character}(object, organism = NULL, format = c("1:1", "long"))
}
\arguments{
\item{object}{Object.}

\item{format}{\code{character(1)}.
Formatting method to apply:
\itemize{
\item \code{"1:1"}: \emph{Recommended.} Return with 1:1 mappings. For Ensembl genes that
don't map 1:1 with Entrez, pick the oldest Entrez identifier. Genes that
don't map to Entrez will contain \code{NA} in \code{entrezId} column.
\item \code{"long"}: Return \code{1:many} in long format.
}}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{...}{Additional arguments.}
}
\value{
\code{Ensembl2Entrez}.
}
\description{
Ensembl-to-Entrez gene identifier mappings
}
\note{
Updated 2022-05-27.
}
\examples{
organism <- "Homo sapiens"

## character ====
## Ensembl-to-Entrez.
genes <- c("ENSG00000000003", "ENSG00000000005")
x <- Ensembl2Entrez(object = genes, organism = organism)
print(x)
}
