% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Ensembl2Ncbi-methods.R
\name{Ensembl2Ncbi}
\alias{Ensembl2Ncbi}
\alias{Ensembl2Ncbi,EnsemblGenes-method}
\alias{Ensembl2Ncbi,GencodeGenes-method}
\alias{Ensembl2Ncbi,character-method}
\title{Ensembl-to-NCBI gene identifier mappings}
\usage{
Ensembl2Ncbi(object, ...)

\S4method{Ensembl2Ncbi}{EnsemblGenes}(object, format = c("1:1", "long"))

\S4method{Ensembl2Ncbi}{GencodeGenes}(object, format = c("1:1", "long"))

\S4method{Ensembl2Ncbi}{character}(object, organism = NULL, format = c("1:1", "long"))
}
\arguments{
\item{object}{Object.}

\item{format}{\code{character(1)}.
Formatting method to apply:
\itemize{
\item \code{"1:1"}: \emph{Recommended.} Return with 1:1 mappings. For Ensembl genes that
don't map 1:1 with NCBI, pick the oldest NCBI identifier. Genes that don't
map to NCBI will contain \code{NA} in \code{ncbiGeneId} column.
\item \code{"long"}: Return \code{1:many} in long format.
}}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{...}{Additional arguments.}
}
\value{
\code{Ensembl2Ncbi}.
}
\description{
Ensembl-to-NCBI gene identifier mappings
}
\note{
Updated 2023-09-16.
}
\examples{
## character ====
x <- Ensembl2Ncbi(
    object = c("ENSG00000000003", "ENSG00000000005"),
    organism = "Homo sapiens"
)
print(x)
}
