% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{EnsemblCds-class}
\alias{EnsemblCds-class}
\title{Ensembl CDS (coding sequence) annotations}
\value{
\code{EnsemblCds}.
}
\description{
Ensembl CDS (coding sequence) annotations
}
\details{
Contains \code{GRangesList} with Ensembl CDS-level annotations grouped by
transcript identifier.
}
\note{
Updated 2026-05-31.
}
