% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{EnsemblExons-class}
\alias{EnsemblExons-class}
\title{Ensembl exon annotations}
\value{
\code{EnsemblExons}.
}
\description{
Ensembl exon annotations
}
\details{
Contains \code{GRanges} with Ensembl exon-level annotations.
}
\note{
Updated 2023-12-20.
}
