% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{EnsemblToNcbi-class}
\alias{EnsemblToNcbi-class}
\title{Ensembl-to-NCBI gene identifier mappings}
\value{
\code{EnsemblToNcbi}.
}
\description{
Ensembl-to-NCBI gene identifier mappings
}
\details{
Contains a \code{DFrame} with \code{"ensemblGeneId"} and \code{"ncbiGeneId"} columns.
}
\note{
Updated 2023-10-04.
}
