% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Entrez2Ensembl-methods.R
\name{Entrez2Ensembl}
\alias{Entrez2Ensembl}
\title{Entrez-to-Ensembl gene identifier mappings}
\usage{
Entrez2Ensembl(object, ...)
}
\value{
\code{Entrez2Ensembl}.
}
\description{
Entrez-to-Ensembl gene identifier mappings
}
\note{
Updated 2021-08-18.
}
\examples{
organism <- "Homo sapiens"

## integer ====
## Entrez-to-Ensembl.
genes <- c(1L, 2L)
x <- Entrez2Ensembl(object = genes, organism = organism)
print(x)
}
