% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{EntrezGeneInfo-class}
\alias{EntrezGeneInfo-class}
\title{NCBI Entrez gene identifier information}
\value{
\code{EntrezGeneInfo}.
}
\description{
NCBI Entrez gene identifier information
}
\note{
Updated 2022-10-24.
}
