% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FlybaseExons-class}
\alias{FlybaseExons-class}
\title{FlyBase exon annotations}
\value{
\code{FlybaseExons}.
}
\description{
FlyBase exon annotations
}
\details{
Contains \code{GRanges} with FlyBase exon-level annotations.
}
\note{
Updated 2023-12-20.
}
