% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FlybaseGenes-class}
\alias{FlybaseGenes-class}
\title{FlyBase gene annotations}
\value{
\code{FlybaseGenes}.
}
\description{
FlyBase gene annotations
}
\details{
Contains \code{GRanges} with FlyBase gene-level annotations.
}
\note{
Updated 2023-09-26.
}
