% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FlyBaseTranscripts-class}
\alias{FlyBaseTranscripts-class}
\title{FlyBase transcript annotations}
\value{
\code{FlyBaseTranscripts}.
}
\description{
FlyBase transcript annotations
}
\details{
Contains \code{GenomicRanges} with FlyBase transcript-level annotations.
}
\note{
Updated 2021-10-14.
}
