% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GencodeExons-class}
\alias{GencodeExons-class}
\title{GENCODE exon annotations}
\value{
\code{GencodeExons}.
}
\description{
GENCODE exon annotations
}
\details{
Contains \code{GRanges} with GENCODE exon-level annotations.
}
\note{
Updated 2023-12-20.
}
