% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GencodeGenes-class}
\alias{GencodeGenes-class}
\title{GENCODE gene annotations}
\value{
\code{GencodeGenes}.
}
\description{
GENCODE gene annotations
}
\details{
Contains a \code{GRanges} with GENCODE gene-level annotations.
}
\note{
Updated 2021-01-25.
}
