% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GeneToSymbol-methods.R
\name{GeneToSymbol}
\alias{GeneToSymbol}
\alias{GeneToSymbol,DFrame-method}
\alias{GeneToSymbol,GRanges-method}
\title{Gene-to-symbol mappings}
\usage{
GeneToSymbol(object, ...)

\S4method{GeneToSymbol}{DFrame}(
  object,
  format = c("makeUnique", "1:1", "unmodified"),
  quiet = FALSE
)

\S4method{GeneToSymbol}{GRanges}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{format}{\code{character(1)}.
Formatting method to apply:
\itemize{
\item \code{"makeUnique"}: \emph{Recommended.} Apply \code{make.unique} to the \code{geneName}
column. Gene names are made unique, while the identifiers remain
unmodified. \code{NA} gene names will be renamed to \code{"unannotated"}.
\item \code{"1:1"}: For gene names that map to multiple gene identifiers, select
only the first annotated gene identifier. Incomplete elements with
\code{NA} gene name will be removed will be removed with an internal
\code{complete.cases} call.
\item \code{"unmodified"}: Return \code{geneId} and \code{geneName} columns unmodified, in
long format. Incomplete elements with \code{NA} gene name will be removed
with an internal \code{complete.cases} call.
}}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}

\item{...}{Arguments pass through to \code{DFrame} method.}
}
\value{
\code{GeneToSymbol}.
}
\description{
Gene-to-symbol mappings
}
\details{
For some organisms, gene identifiers and gene names do not map 1:1 (e.g.
\emph{Homo sapiens} and \emph{Mus musculus}). Refer to the \code{format} argument
here in the documentation for approaches that deal with this issue.
}
\note{
Updated 2023-09-16.
}
\examples{
data(GRanges, package = "AcidTest")

## DFrame ====
df <- S4Vectors::DataFrame(
    "geneId" = c(
        "ENSG00000228572.7",
        "ENSG00000182378.14"
    ),
    "geneName" = c(
        "AL954722.1",
        "PLCXD1"
    )
)
x <- GeneToSymbol(df)
print(x)

## GRanges ====
object <- GRanges
x <- GeneToSymbol(object)
}
\seealso{
\code{\link[=makeGeneToSymbol]{makeGeneToSymbol()}}.
}
