% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnsDb.R
\name{getEnsDb}
\alias{getEnsDb}
\title{Get EnsDb from Bioconductor}
\usage{
getEnsDb(organism, genomeBuild = NULL, release = NULL)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
\emph{Note:} don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{release}{\code{integer(1)}.
Ensembl release version (e.g. \code{100}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}
}
\value{
\code{EnsDb}.
}
\description{
Get EnsDb from Bioconductor
}
\details{
Primarily fetches \code{EnsDb} objects from AnnotationHub.
For legacy GRCh37 genome build, loads \code{EnsDb.Hsapiens.v75} data package.
}
\note{
Updated 2023-12-20.
}
\section{UCSC genome build remapping}{


Remaps UCSC genome build to Ensembl automatically, if necessary.
For example, \code{"hg38"} remaps to \code{"GRCh38"} and "hg19" to \code{"GRCh37"}.
}

\examples{
## Current genome build and release.
object <- getEnsDb(organism = "Homo sapiens")
print(object)

## Legacy GRCh37 genome build.
if (requireNamespace("EnsDb.Hsapiens.v75", quietly = TRUE)) {
    object <- getEnsDb(organism = "Homo sapiens", genomeBuild = "GRCh37")
    print(object)
}
}
