% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGFFDirectives.R
\name{getGFFDirectives}
\alias{getGFFDirectives}
\title{Get the directives from a GFF file}
\usage{
getGFFDirectives(file, nMax = Inf)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}

\item{nMax}{\code{integer(1)}. or \code{Inf}.
Maximum number (i.e. of lines) to include.
Use \code{Inf} to not apply a limit.}
}
\value{
\code{DataFrame} or \code{NULL}.
}
\description{
Get the directives from a GFF file
}
\details{
Matches lines beginning with \verb{#!<key> <value>} or \verb{##<key>: <value>}
}
\note{
Updated 2021-01-21.
}
\section{GFF3}{

Lines beginning with '##' are directives (sometimes called pragmas or
meta-data) and provide meta-information about the document as a whole. Blank
lines should be ignored by parsers and lines beginning with a single '#' are
used for human-readable comments and can be ignored by parsers. End-of-line
comments (comments preceded by # at the end of and on the same line as a
feature or directive line) are not allowed.
}

\examples{
url <- pasteURL(
    "ftp.ensembl.org",
    "pub",
    "release-102",
    "gtf",
    "homo_sapiens",
    "Homo_sapiens.GRCh38.102.gtf.gz",
    protocol = "ftp"
)
df <- getGFFMetadata(url)
print(df)
}
\seealso{
\itemize{
\item https://github.com/The-Sequence-Ontology/Specifications/blob/master/gff3.md
}
}
