% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGFFMetadata.R
\name{getGFFMetadata}
\alias{getGFFMetadata}
\title{Get the metadata (directives) from a GFF file}
\usage{
getGFFMetadata(file, nMax = Inf)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}
}
\value{
\code{DataFrame} or \code{NULL}.
}
\description{
Get the metadata (directives) from a GFF file
}
\details{
Matches lines beginning with \verb{#!<key> <value>} or \verb{##<key>: <value>}
}
\note{
Updated 2021-01-20.
}
\section{GFF3}{

Lines beginning with '##' are directives (sometimes called pragmas or
meta-data) and provide meta-information about the document as a whole. Blank
lines should be ignored by parsers and lines beginning with a single '#' are
used for human-readable comments and can be ignored by parsers. End-of-line
comments (comments preceded by # at the end of and on the same line as a
feature or directive line) are not allowed.
}

\examples{
url <- pasteURL(
    "ftp.ensembl.org",
    "pub",
    "release-102",
    "gtf",
    "homo_sapiens",
    "Homo_sapiens.GRCh38.102.gtf.gz",
    protocol = "ftp"
)
df <- getGFFMetadata(url)
print(df)
}
\seealso{
\itemize{
\item https://github.com/The-Sequence-Ontology/Specifications/blob/master/gff3.md
}
}
