% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRefSeqAssemblySummary.R
\name{getRefSeqAssemblySummary}
\alias{getRefSeqAssemblySummary}
\title{Get the RefSeq assembly metadata}
\usage{
getRefSeqAssemblySummary(file)
}
\arguments{
\item{file}{\code{character(1)}.
RefSeq assembly summary file or URL.}
}
\value{
Named \code{character}.
}
\description{
Get the RefSeq assembly metadata
}
\note{
Updated 2021-01-14.
}
\section{Stable reference assembly}{

The latest reference assembly, linked under the \verb{reference/}
subdirectory changes over time and is not considered "stable". For improved
reproduciblity, track a reference one version behind
(e.g. use "GCF_000001405.39_GRCh38.p12" instead of current
"GCF_000001405.39_GRCh38.p13" build).

This approach to maintaining a current reference build differs on NCBI RefSeq
compared to other sources such as Ensembl and GENCODE, which track a stable
release as the latest "reference" assembly.
}

\examples{
file <- pasteURL(
    "ftp.ncbi.nlm.nih.gov",
    "genomes",
    "refseq",
    "vertebrate_mammalian",
    "Homo_sapiens",
    "assembly_summary.txt",
    protocol = "ftp"
)
x <- getRefSeqAssemblySummary(file)
names(x)
}
\seealso{
\itemize{
\item \href{ftp://ftp.ncbi.nlm.nih.gov/genomes/README_assembly_summary.txt}{File format details}.
}
}
