% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRefSeqSeqinfo.R
\name{getRefSeqSeqinfo}
\alias{getRefSeqSeqinfo}
\title{Get RefSeq genome assembly seqinfo}
\usage{
getRefSeqSeqinfo(file)
}
\arguments{
\item{file}{\code{character(1)}.
RefSeq assembly report file or URL.}
}
\value{
\code{Seqinfo}.
}
\description{
Parse the assembly report file to get \code{seqlengths} per chromosome.
}
\note{
Updated 2021-01-14.
}
\examples{
file <- pasteURL(
    "ftp.ncbi.nlm.nih.gov",
    "genomes",
    "refseq",
    "vertebrate_mammalian",
    "Homo_sapiens",
    "all_assembly_versions",
    "GCF_000001405.38_GRCh38.p12",
    "GCF_000001405.38_GRCh38.p12_assembly_report.txt",
    protocol = "ftp"
)
seqinfo <- getRefSeqSeqinfo(file)
print(seqinfo)
}
\seealso{
\itemize{
\item \code{tximeta:::gtf2RefSeq}.
}
}
