% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goTermsPerGeneName.R
\name{goTermsPerGeneName}
\alias{goTermsPerGeneName}
\title{Import gene ontology (GO) terms per gene name}
\usage{
goTermsPerGeneName(
  organism,
  geneNames = NULL,
  format = c("long", "split", "nested")
)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{geneNames}{\code{character} or \code{NULL}.
Return specific genes, or when \code{NULL}, return all genes.}

\item{format}{\code{character(1)}.
Return GO terms in \code{"long"}, \code{"split"}, or \code{"nested"} formats.
Processing GO terms into nested format is CPU intensive, and is recommended
to set \code{geneNames} in this case.}
}
\value{
Varies, depending on \code{format} argument.
\itemize{
\item \code{"long"}: \code{DFrame} with \code{"geneName"}, \code{"goCategory"}, \code{"goId"}, and
\code{"goName"} columns.
\item \code{"split"}: \code{SplitDFrameList}: Split by \code{geneName} column. Contains
\code{"geneName"}, \code{"goCategory"}, \code{"goId"}, and \code{"goName"} columns.
\item \code{"nested"}: \code{DFrame} with \code{"geneName"}, \code{"goBp"}, \code{"goCc"}, and \code{"goMf"}
columns. Does not contain any duplicates in \code{"geneName"} column.
}
}
\description{
Import gene ontology (GO) terms per gene name
}
\note{
Updated 2023-12-15.
}
\examples{
object <- goTermsPerGeneName(organism = "Homo sapiens", format = "long")
print(object)
}
\seealso{
\itemize{
\item http://current.geneontology.org/products/pages/downloads.html
\item https://www.ebi.ac.uk/GOA/
\item https://geneontology.org/docs/go-annotation-file-gaf-format-2.2/
\item https://ftp.ncbi.nlm.nih.gov/gene/DATA/gene2go.gz
\item Python goatools package.
}
}
