% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Hgnc-class}
\alias{Hgnc-class}
\title{Human Genome Organization (HUGO) Gene Nomenclature Committee (HGNC) metadata}
\value{
\code{Hgnc}.
}
\description{
Human Genome Organization (HUGO) Gene Nomenclature Committee (HGNC) metadata
}
\note{
Updated 2025-03-24.
}
