% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hgnc.R
\name{Hgnc}
\alias{Hgnc}
\title{Import Human Genome Organization (HUGO) Gene Nomenclature Committee (HGNC)
metadata}
\usage{
Hgnc()
}
\value{
\code{Hgnc}.
}
\description{
Import Human Genome Organization (HUGO) Gene Nomenclature Committee (HGNC)
metadata
}
\note{
Updated 2023-11-21.
}
\examples{
object <- Hgnc()
print(object)
}
\seealso{
\itemize{
\item https://www.genenames.org/
\item https://www.genenames.org/download/statistics-and-files/
}
}
