% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanToMouse.R
\name{HumanToMouse}
\alias{HumanToMouse}
\title{Import human-to-mouse gene mappings}
\usage{
HumanToMouse()
}
\value{
\code{HumanToMouse}.
}
\description{
Import human-to-mouse gene mappings
}
\note{
Updated 2023-09-26.
}
\examples{
h2m <- HumanToMouse()
print(h2m)
}
\seealso{
\itemize{
\item https://www.informatics.jax.org/
\item https://www.biostars.org/p/9567892/
}
}
