% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importEnsemblToNcbiGeneMap.R
\name{importEnsemblToNcbiGeneMap}
\alias{importEnsemblToNcbiGeneMap}
\title{Import Ensembl-to-NCBI gene identifier mappings}
\usage{
importEnsemblToNcbiGeneMap(
  organism,
  genomeBuild = NULL,
  release = NULL,
  uniqueOnly = TRUE
)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
\emph{Note:} don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{release}{\code{integer(1)}.
Ensembl release version (e.g. \code{100}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}

\item{uniqueOnly}{\code{logical(1)}.
Resolve duplicates and only return unique 1:1 mappings.}
}
\value{
\code{DFrame}.
}
\description{
Import Ensembl-to-NCBI gene identifier mappings
}
\note{
Updated 2025-04-15.
}
\examples{
df <- importEnsemblToNcbiGeneMap(organism = "Homo sapiens")
print(df)
}
