% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{JaxHumanToMouse-class}
\alias{JaxHumanToMouse-class}
\title{Jackson Laboratory (JAX) human-to-mouse gene mappings}
\value{
\code{JaxHumanToMouse}.
}
\description{
Jackson Laboratory (JAX) human-to-mouse gene mappings
}
\details{
Contains a \code{DFrame} with \code{"dbClassKey"}, \code{"humanGeneName"}, \code{"humanHgncId"},
\code{"humanNcbiGeneId"}, \code{"humanOmimGeneId"}, \code{"mouseGeneName"}, \code{"mouseMgiId"},
\code{"mouseNcbiGeneId"} columns.
}
\note{
Updated 2023-11-22.
}
