% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaxHumanToMouse.R
\name{JaxHumanToMouse}
\alias{JaxHumanToMouse}
\title{Import human-to-mouse gene mappings from the Jackson Laboratory (JAX) Mouse
Genome Informatics (MGI) website}
\usage{
JaxHumanToMouse(unique = TRUE)
}
\arguments{
\item{unique}{\code{logical(1)}.
Only return unique 1:1 mappings between human and mouse.
Recommended by default.}
}
\value{
\code{JaxHumanToMouse}.
}
\description{
Import human-to-mouse gene mappings from the Jackson Laboratory (JAX) Mouse
Genome Informatics (MGI) website
}
\note{
Updated 2023-10-12.
}
\examples{
object <- JaxHumanToMouse()
print(object)
}
\seealso{
\itemize{
\item https://www.informatics.jax.org/
\item https://www.biostars.org/p/9567892/
}
}
