% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEnsDbFromGff.R
\name{makeEnsDbFromGff}
\alias{makeEnsDbFromGff}
\title{Make EnsDb object from a GFF/GTF file}
\usage{
makeEnsDbFromGff(file)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}
}
\value{
\code{EnsDb}.
}
\description{
Wrapper for ensembldb importer functions.
}
\note{
Updated 2023-04-26.
}
\examples{
## > file <- AcidBase::pasteUrl(
## >     "ftp.ensembl.org",
## >     "pub",
## >     "release-102",
## >     "gtf",
## >     "homo_sapiens",
## >     "Homo_sapiens.GRCh38.102.gtf.gz",
## >     protocol = "ftp"
## > )
## > edb <- makeEnsDbFromGff(file)
## > print(edb)
}
\seealso{
\itemize{
\item \code{ensembldb::ensDbFromGff()}.
\item \code{ensembldb::ensDbFromGtf()}.
}
}
