% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGeneToSymbol.R
\name{makeGeneToSymbol}
\alias{makeGeneToSymbol}
\alias{makeGeneToSymbolFromEnsembl}
\alias{makeGeneToSymbolFromEnsDb}
\alias{makeGeneToSymbolFromGFF}
\title{Make a GeneToSymbol object}
\usage{
makeGeneToSymbolFromEnsembl(
  organism,
  genomeBuild = NULL,
  release = NULL,
  ignoreVersion = TRUE,
  format = c("makeUnique", "1:1", "unmodified")
)

makeGeneToSymbolFromEnsDb(
  object,
  ignoreVersion = TRUE,
  format = c("makeUnique", "1:1", "unmodified")
)

makeGeneToSymbolFromGFF(
  file,
  ignoreVersion = TRUE,
  format = c("makeUnique", "1:1", "unmodified")
)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
\emph{Note:} don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{release}{\code{integer(1)}.
Ensembl release version (e.g. \code{100}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}

\item{ignoreVersion}{\code{logical(1)}.
Ignore identifier (e.g. transcript, gene) versions.
When applicable, the identifier containing version numbers will be stored
in \code{txIdVersion} and \code{geneIdVersion}, and the variants without versions
will be stored in \code{txId}, \code{txIdNoVersion}, \code{geneId}, and \code{geneIdNoVersion}.}

\item{format}{\code{character(1)}.
Formatting method to apply:
\itemize{
\item \code{"makeUnique"}: \emph{Recommended.} Apply \code{make.unique} to the \code{geneName}
column. Gene names are made unique, while the identifiers remain
unmodified. \code{NA} gene names will be renamed to \code{"unannotated"}.
\item \code{"1:1"}: For gene names that map to multiple gene identifiers, select
only the first annotated gene identifier. Incomplete elements with
\code{NA} gene name will be removed will be removed with an internal
\code{complete.cases} call.
\item \code{"unmodified"}: Return \code{geneId} and \code{geneName} columns unmodified, in
long format. Incomplete elements with \code{NA} gene name will be removed
with an internal \code{complete.cases} call.
}}

\item{object}{Object.}

\item{file}{\code{character(1)}.
File path.}
}
\value{
\code{GeneToSymbol}.
}
\description{
Make a GeneToSymbol object
}
\section{Functions}{
\itemize{
\item \code{makeGeneToSymbolFromEnsembl()}: Make a \code{GeneToSymbol} object from Ensembl using
an AnnotationHub lookup.

\item \code{makeGeneToSymbolFromEnsDb()}: Make a \code{GeneToSymbol} object from an \code{EnsDb}
object or annotation package.

\item \code{makeGeneToSymbolFromGFF()}: Make a \code{GeneToSymbol} object from a GFF file.

}}
\note{
Updated 2021-08-03.
}
\section{GFF/GTF file}{


Remote URLs and compressed files are supported.
}

\examples{
## makeGeneToSymbolFromEnsembl ====
x <- makeGeneToSymbolFromEnsembl(
    organism = "Homo sapiens",
    ignoreVersion = FALSE
)
print(x)

## makeTxToGeneFromEnsDb ====
if (goalie::isInstalled("EnsDb.Hsapiens.v75")) {
    x <- makeGeneToSymbolFromEnsDb("EnsDb.Hsapiens.v75")
    print(x)
}

## makeGeneToSymbolFromGFF ====
file <- AcidBase::pasteURL(
    "ftp.ensembl.org",
    "pub",
    "release-102",
    "gtf",
    "homo_sapiens",
    "Homo_sapiens.GRCh38.102.gtf.gz",
    protocol = "https"
)
x <- makeGeneToSymbolFromGFF(
    file = file,
    ignoreVersion = FALSE
)
print(x)
}
