% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesFromEnsDb.R
\name{makeGRangesFromEnsDb}
\alias{makeGRangesFromEnsDb}
\title{Make GRanges from EnsDb object}
\usage{
makeGRangesFromEnsDb(
  object,
  level = c("genes", "transcripts"),
  ignoreVersion = TRUE,
  broadClass = TRUE,
  synonyms = FALSE
)
}
\arguments{
\item{object}{Object.}

\item{level}{\code{character(1)}.
Return as genes or transcripts.}

\item{broadClass}{\code{logical(1)}.
Include broad class definitions.}

\item{synonyms}{\code{logical(1)}.
Include gene synonyms.
Queries the Ensembl web server, and is CPU intensive.}
}
\value{
\code{GRanges}.
}
\description{
Make GRanges from EnsDb object
}
\details{
Use specific \code{EnsDb} object as annotation source.
Alternatively, can pass in an EnsDb package name as a \code{character(1)}.
}
\note{
Updated 2021-01-10.
}
\examples{
if ("EnsDb.Hsapiens.v75" \%in\% rownames(installed.packages())) {
    x <- makeGRangesFromEnsDb("EnsDb.Hsapiens.v75")
}
}
