% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesFromTxDb.R
\name{makeGRangesFromTxDb}
\alias{makeGRangesFromTxDb}
\title{Make GenomicRanges from TxDb object}
\usage{
makeGRangesFromTxDb(
  object,
  level = c("transcripts", "genes", "exons", "cds"),
  ignoreVersion = TRUE,
  synonyms = FALSE
)
}
\arguments{
\item{object}{Object.}

\item{level}{\code{character(1)}.
Return as genes or transcripts.}

\item{ignoreVersion}{\code{logical(1)}.
Ignore identifier (e.g. transcript, gene) versions.
When applicable, the identifier containing version numbers will be stored
in \code{txIdVersion} and \code{geneIdVersion}, and the variants without versions
will be stored in \code{txId}, \code{txIdNoVersion}, \code{geneId}, and \code{geneIdNoVersion}.}

\item{synonyms}{\code{logical(1)}.
Include gene synonyms.
Queries the Ensembl web server, and is CPU intensive.}
}
\value{
\code{GenomicRanges}.
}
\description{
Make GenomicRanges from TxDb object
}
\note{
Updated 2021-04-27.
}
\examples{
file <- pasteURL(
    "ftp.ensembl.org",
    "pub",
    "release-102",
    "gtf",
    "homo_sapiens",
    "Homo_sapiens.GRCh38.102.gtf.gz",
    protocol = "ftp"
)
txdb <- AcidGenomes::makeTxDbFromGFF(file)
gr <- makeGRangesFromTxDb(object = txdb, ignoreVersion = FALSE)
print(gr)
}
