% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTx2Gene.R
\name{makeTx2Gene}
\alias{makeTx2Gene}
\alias{makeTx2GeneFromEnsembl}
\alias{makeTx2GeneFromEnsDb}
\alias{makeTx2GeneFromGFF}
\title{Make a Tx2Gene object}
\usage{
makeTx2GeneFromEnsembl(
  organism,
  genomeBuild = NULL,
  release = NULL,
  ignoreVersion = FALSE
)

makeTx2GeneFromEnsDb(object, ignoreVersion = FALSE)

makeTx2GeneFromGFF(file, ignoreVersion = FALSE)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
\emph{Note:} don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{release}{\code{integer(1)}.
Ensembl release version (e.g. \code{100}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}

\item{ignoreVersion}{\code{logical(1)}.
Ignore identifier (e.g. transcript, gene) versions.
When applicable, the identifier containing version numbers will be stored
in \code{txIdVersion} and \code{geneIdVersion}, and the variants without versions
will be stored in \code{txId}, \code{txIdNoVersion}, \code{geneId}, and \code{geneIdNoVersion}.}

\item{object}{Object.}

\item{file}{\code{character(1)}.
File path.}
}
\value{
\code{Tx2Gene}.
}
\description{
Make a Tx2Gene object
}
\note{
Updated 2023-07-31.
}
\section{GFF/GTF file}{


Remote URLs and compressed files are supported.
}

\examples{
## makeTx2GeneFromEnsembl ====
x <- makeTx2GeneFromEnsembl(organism = "Homo sapiens")
print(x)

## makeTx2GeneFromEnsDb ====
if (goalie::isInstalled("EnsDb.Hsapiens.v75")) {
    x <- makeTx2GeneFromEnsDb(object = "EnsDb.Hsapiens.v75")
    print(x)
}

## makeTx2GeneFromGFF ====
file <- AcidBase::pasteURL(
    "ftp.ensembl.org",
    "pub",
    "release-102",
    "gtf",
    "homo_sapiens",
    "Homo_sapiens.GRCh38.102.gtf.gz",
    protocol = "ftp"
)
x <- makeTx2GeneFromGFF(file = file)
print(x)
}
