% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTx2GeneFromFASTA.R
\name{makeTx2GeneFromFASTA}
\alias{makeTx2GeneFromFASTA}
\title{Make a Tx2Gene object from transcriptome FASTA}
\usage{
makeTx2GeneFromFASTA(file)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}
}
\value{
\code{Tx2Gene}.
}
\description{
Make a Tx2Gene object from transcriptome FASTA
}
\note{
RefSeq transcript FASTA
(e.g. "GCF_000001405.39_GRCh38.p13_rna.fna.gz") doesn't contain gene
identifiers, and is not supported.

Updated 2021-01-29.
}
\examples{
## Ensembl ====
file <- AcidBase::pasteURL(
    "ftp.ensembl.org",
    "pub",
    "release-102",
    "fasta",
    "homo_sapiens",
    "cdna",
    "Homo_sapiens.GRCh38.cdna.all.fa.gz",
    protocol = "ftp"
)
t2g <- makeTx2GeneFromFASTA(file)
print(t2g)

## GENCODE ====
## > file <- AcidBase::pasteURL(
## >     "ftp.ebi.ac.uk",
## >     "pub",
## >     "databases",
## >     "gencode",
## >     "Gencode_human",
## >     "release_32",
## >     "gencode.v32.transcripts.fa.gz",
## >     protocol = "ftp"
## > )
## > t2g <- makeTx2GeneFromFASTA(file)
## > print(t2g)

## FlyBase ====
## > file <- AcidBase::pasteURL(
## >     "ftp.flybase.net",
## >     "releases",
## >     "FB2019_05",
## >     "dmel_r6.30",
## >     "fasta",
## >     "dmel-all-transcript-r6.30.fasta.gz",
## >     protocol = "ftp"
## > )
## > t2g <- makeTx2GeneFromFASTA(file)
## > print(t2g)

## WormBase ====
## > file <- AcidBase::pasteURL(
## >     "ftp.wormbase.org",
## >     "pub",
## >     "wormbase",
## >     "releases",
## >     "WS272",
## >     "species",
## >     "c_elegans",
## >     "PRJNA13758",
## >     "c_elegans.PRJNA13758.WS272.mRNA_transcripts.fa.gz",
## >     protocol = "ftp"
## > )
## > t2g <- makeTx2GeneFromFASTA(file)
## > print(t2g)
}
