% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTxDbFromGFF.R
\name{makeTxDbFromGFF}
\alias{makeTxDbFromGFF}
\alias{makeTxDbFromGTF}
\title{Make TxDb from GRanges}
\usage{
makeTxDbFromGFF(
  file,
  seqinfo,
  organism = NULL,
  genomeBuild = NULL,
  source = file
)

makeTxDbFromGTF(
  file,
  seqinfo,
  organism = NULL,
  genomeBuild = NULL,
  source = file
)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}

\item{seqinfo}{\code{Seqinfo}.
Information about the chromosomes.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
Note: don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{source}{\code{character(1)}.
Description of data source.
Defaults to file name.}
}
\value{
\code{TxDb}.
}
\description{
Make TxDb from GRanges
}
\details{
This step can be noisy and generate expected warnings, which are suppressed:\preformatted{some exons are linked to transcripts not found in the file
The following orphan exon were dropped
The following orphan CDS were dropped
}
}
\section{Functions}{
\itemize{
\item \code{makeTxDbFromGFF}: Primary function.

\item \code{makeTxDbFromGTF}: Alias for GTF files.
}}

\note{
Updated 2021-01-14.
}
\examples{
file <- pasteURL(AcidGenomesTestsURL, "ensembl.gtf")
txdb <- makeTxDbFromGFF(file)
print(txdb)
}
\seealso{
\itemize{
\item \code{GenomicFeatures::makeTxDbFromGFF()}.
\item \code{GenomicFeatures::supportedMiRBaseBuildValues()}.
Note that \emph{Homo sapiens} GRCh38 isn't currently supported in mirbase.db.
\item \href{https://bioconductor.org/packages/TxDb.Hsapiens.UCSC.hg38.knownGene/}{TxDb.Hsapiens.UCSC.hg38.knownGene}.
\item https://stackoverflow.com/questions/38603668
\item https://stackoverflow.com/questions/16517795
\item https://github.com/Bioconductor/GenomicFeatures/blob/master/R/makeTxDbFromGRanges.R
\item https://github.com/Bioconductor/GenomicFeatures/issues/26
}
}
