% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTxToGene.R
\name{makeTxToGene}
\alias{makeTxToGene}
\alias{makeTxToGeneFromEnsembl}
\alias{makeTxToGeneFromEnsDb}
\alias{makeTxToGeneFromGFF}
\title{Make a TxToGene object}
\usage{
makeTxToGeneFromEnsembl(
  organism,
  genomeBuild = NULL,
  release = NULL,
  ignoreVersion = FALSE
)

makeTxToGeneFromEnsDb(object, ignoreVersion = FALSE)

makeTxToGeneFromGFF(file, ignoreVersion = FALSE)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
\emph{Note:} don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{release}{\code{integer(1)}.
Ensembl release version (e.g. \code{100}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}

\item{ignoreVersion}{\code{logical(1)}.
Ignore identifier (e.g. transcript, gene) versions.
When applicable, the identifier containing version numbers will be stored
in \code{txIdVersion} and \code{geneIdVersion}, and the variants without versions
will be stored in \code{txId}, \code{txIdNoVersion}, \code{geneId}, and \code{geneIdNoVersion}.}

\item{object}{Object.}

\item{file}{\code{character(1)}.
File path.}
}
\value{
\code{TxToGene}.
}
\description{
Make a TxToGene object
}
\note{
Updated 2023-07-31.
}
\section{GFF/GTF file}{


Remote URLs and compressed files are supported.
}

\examples{
## makeTxToGeneFromEnsembl ====
x <- makeTxToGeneFromEnsembl(organism = "Homo sapiens")
print(x)

## makeTxToGeneFromEnsDb ====
if (goalie::isInstalled("EnsDb.Hsapiens.v75")) {
    x <- makeTxToGeneFromEnsDb(object = "EnsDb.Hsapiens.v75")
    print(x)
}

## makeTxToGeneFromGFF ====
file <- AcidBase::pasteURL(
    "ftp.ensembl.org",
    "pub",
    "release-102",
    "gtf",
    "homo_sapiens",
    "Homo_sapiens.GRCh38.102.gtf.gz",
    protocol = "https"
)
x <- makeTxToGeneFromGFF(file = file)
print(x)
}
