% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapEnsemblReleaseToUrl.R
\name{mapEnsemblReleaseToUrl}
\alias{mapEnsemblReleaseToUrl}
\title{Map Ensembl release to archive URL.}
\usage{
mapEnsemblReleaseToUrl(release)
}
\arguments{
\item{release}{\code{integer(1)} or \code{character(1)}.
Ensembl release (e.g. \code{100}).}
}
\value{
\code{character(1)}.
Ensembl release URL.
}
\description{
Map Ensembl release to archive URL.
}
\details{
Requires the rvest package to be installed.
}
\note{
Updated 2023-09-27.
}
\examples{
try({
    mapEnsemblReleaseToUrl(release = 100L)
})
}
\seealso{
\itemize{
\item \code{biomaRt::listEnsemblArchives()}.
}
}
