% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapEnsemblToGencode.R
\name{mapEnsemblToGencode}
\alias{mapEnsemblToGencode}
\title{Map Ensembl release to GENCODE}
\usage{
mapEnsemblToGencode(release, organism = "Homo sapiens")
}
\arguments{
\item{release}{\code{integer(1)}.
Ensembl release (e.g. \code{110}).}

\item{organism}{\code{character(1)}.
Scientific name of the organism.
Currently only \code{"Homo sapiens"} and \code{"Mus musculus"} are supported.}
}
\value{
\code{character(1)}.
Corresponding GENCODE release identifier (e.g. \code{"44"} for human or
\code{"M33"} for mouse).
}
\description{
Map Ensembl release to GENCODE
}
\note{
Updated 2026-05-31.
}
\examples{
## Homo sapiens.
mapEnsemblToGencode(110L, organism = "Homo sapiens")
## Mus musculus.
mapEnsemblToGencode(110L, organism = "Mus musculus")
}
\seealso{
\itemize{
\item https://www.gencodegenes.org/human/releases.html
\item https://www.gencodegenes.org/mouse/releases.html
\item \code{\link[=mapGencodeToEnsembl]{mapGencodeToEnsembl()}}.
}
}
