% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGencodeToEnsembl.R
\name{mapGencodeToEnsembl}
\alias{mapGencodeToEnsembl}
\title{Map GENCODE release to Ensembl}
\usage{
mapGencodeToEnsembl(release)
}
\arguments{
\item{release}{\code{integer(1)}.
Human (e.g. \code{42}) or mouse (e.g. \code{"M21"}) GENCODE release.}
}
\value{
\code{integer(1)}.
Ensembl release.
}
\description{
Map GENCODE release to Ensembl
}
\note{
Updated 2023-01-30.
}
\examples{
## Homo sapiens.
mapGencodeToEnsembl(42L)
## Mus musculus.
mapGencodeToEnsembl("M21L")
}
\seealso{
\itemize{
\item https://www.gencodegenes.org/human/releases.html
\item https://www.gencodegenes.org/mouse/releases.html
\item https://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/_README.TXT
\item https://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_mouse/_README.TXT
}
}
