% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGeneNamesToHGNC.R
\name{mapGeneNamesToHGNC}
\alias{mapGeneNamesToHGNC}
\title{Map gene names (symbols) to HGNC identifiers}
\usage{
mapGeneNamesToHGNC(hgnc, genes)
}
\arguments{
\item{hgnc}{\code{HGNC}.}

\item{genes}{\code{character}.
Gene names (e.g. \code{"TP53"}).}
}
\description{
Map gene names (symbols) to HGNC identifiers
}
\note{
Updated 2023-03-02.
}
