% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGeneNamesToNCBI.R
\name{mapGeneNamesToNCBI}
\alias{mapGeneNamesToNCBI}
\title{Map gene names to NCBI}
\usage{
mapGeneNamesToNCBI(genes, organism, ncbi)
}
\arguments{
\item{genes}{\code{character}.
Gene names (e.g. \code{"TUT4"}).}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{ncbi}{\code{NcbiGeneIfo} or \code{NULL}.
If \code{NULL}, NCBI annotations will be downloaded automatically.}
}
\description{
Map gene names to NCBI
}
\note{
Updated 2023-03-03.
}
\examples{
x <- mapGeneNamesToNCBI(
    genes = c("TUT4", "ZCCHC11", "TENT3A"),
    organism = "Homo sapiens"
)
}
