% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGenomeBuild.R
\name{mapGenomeBuild}
\alias{mapGenomeBuild}
\alias{mapUCSCBuildToEnsembl}
\alias{mapEnsemblBuildToUCSC}
\title{Interconvert between EBI (Ensembl), NCBI, and UCSC genome build names}
\usage{
mapUCSCBuildToEnsembl(...)

mapEnsemblBuildToUCSC(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{character}. Ensembl genome build as the value, UCSC build as the
name. Stops on match failure.
}
\description{
Interconvert between EBI (Ensembl), NCBI, and UCSC genome build names
}
\note{
Updated 2021-01-20.
}
\examples{
from <- c("hg19", "hg38")
to <- mapUCSCBuildToEnsembl(from)
print(to)
}
\seealso{
\itemize{
\item \href{https://genome.ucsc.edu/cgi-bin/hgGateway}{UCSC hgGateway}.
\item \href{https://goo.gl/164J2P}{bcbio genome recipes}.
}
}
