% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapHumanOrthologs.R
\name{mapHumanOrthologs}
\alias{mapHumanOrthologs}
\title{Map input to human gene orthologs}
\usage{
mapHumanOrthologs(genes, organism = NULL, ensemblRelease = NULL)
}
\arguments{
\item{genes}{\code{character}.
Gene identifiers.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{ensemblRelease}{\code{integer(1)}.
Ensembl release version (e.g. \code{100}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}
}
\value{
\code{DFrame}.
Data frame containing mapping columns:
\itemize{
\item \code{geneId}
\item \code{geneName}
\item \code{humanGeneId}
\item \code{humanGeneName}
}
}
\description{
Map input to human gene orthologs
}
\details{
Genes with identifier versions (e.g. "ENSMUSG00000000001.5") are not
currently supported.
}
\note{
Updated 2023-04-26.
}
\examples{
## Expecting ENSMUSG00000000001, ENSMUSG00000000003 to not match here.
genes <- c(
    "ENSMUSG00000000001", "ENSMUSG00000000003",
    "ENSMUSG00000000028", "ENSMUSG00000000031",
    "ENSMUSG00000000037", "ENSMUSG00000000049"
)
## Protect against Ensembl timeouts causing build checks to fail.
if (goalie::isAnExistingURL("https://ensembl.org")) {
    try({
        x <- mapHumanOrthologs(genes = genes, ensemblRelease = 87L)
        print(x)
    })
}
}
\seealso{
\itemize{
\item \code{biomaRt::listEnsemblArchives()}.
\item \code{biomaRt::listMarts()}.
\item \code{biomaRt::useMart()}.
}
}
