% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapRnacentral.R
\name{mapRnacentral}
\alias{mapRnacentral}
\title{Import RNAcentral identifier mappings}
\usage{
mapRnacentral(organism, database)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{database}{\code{character(1)}.
Database name. Currently supported: \code{"Ensembl"}, \code{"GENCODE"}, \code{"HGNC"},
\code{"MGI"}, and \code{"RefSeq"}. HGNC database is \emph{Homo sapiens} specific, and MGI is
\emph{Mus musculus} specific.}
}
\value{
\code{DFrame}.
Contains RNAcentral to specified database RNA transcript identifier mappings.
Returns \code{"rnacentralId"}, \code{"database"}, \code{"txId"}, \code{"taxId"}, \code{"txBiotype"},
and \code{"geneId"} columns.
}
\description{
Import RNAcentral identifier mappings
}
\note{
Updated 2023-12-13.
}
\examples{
## Get RNAcentral to RefSeq identifier mappings.
object <- mapRnacentral(organism = "Homo sapiens", database = "refseq")
print(object)
}
\seealso{
\itemize{
\item https://rnacentral.org/
\item https://ftp.ebi.ac.uk/pub/databases/RNAcentral/
}
}
