% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapUCSCBuildToEnsembl.R
\name{mapUCSCBuildToEnsembl}
\alias{mapUCSCBuildToEnsembl}
\title{Map UCSC genome build to Ensembl}
\usage{
mapUCSCBuildToEnsembl(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{character}. Ensembl genome build as the value, UCSC build as the
name. Stops on match failure.
}
\description{
Map UCSC genome build to Ensembl
}
\note{
Updated 2020-10-06.
}
\examples{
from <- c("hg19", "hg38")
to <- mapUCSCBuildToEnsembl(from)
print(to)
}
\seealso{
\itemize{
\item \href{https://genome.ucsc.edu/cgi-bin/hgGateway}{UCSC hgGateway}
\item \href{https://goo.gl/164J2P}{bcbio genome recipes}
}
}
