% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapUCSCBuildToNCBI.R
\name{mapUCSCBuildToNCBI}
\alias{mapUCSCBuildToNCBI}
\alias{mapUCSCBuildToEnsembl}
\title{Map UCSC genome build to NCBI}
\usage{
mapUCSCBuildToNCBI(object)

mapUCSCBuildToEnsembl(...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}. Ensembl genome build as the value, UCSC build as the
name. Stops on match failure.
}
\description{
Map UCSC genome build to NCBI
}
\note{
Updated 2021-01-20.
}
\examples{
from <- c("hg19", "hg38")
to <- mapUCSCBuildToEnsembl(from)
print(to)
}
\seealso{
\itemize{
\item \href{https://genome.ucsc.edu/cgi-bin/hgGateway}{UCSC hgGateway}
\item \href{https://goo.gl/164J2P}{bcbio genome recipes}
}
}
