% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Ncbi2Ensembl-class}
\alias{Ncbi2Ensembl-class}
\title{NCBI-to-Ensembl gene identifier mappings}
\value{
\code{Ncbi2Ensembl}.
}
\description{
NCBI-to-Ensembl gene identifier mappings
}
\details{
Contains a \code{DFrame} with \code{ncbiGeneId} and \code{ensemblGeneId} columns.
}
\note{
Updated 2023-09-16.
}
