% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/NcbiToEnsembl-methods.R
\name{NcbiToEnsembl}
\alias{NcbiToEnsembl}
\alias{NcbiToEnsembl,integer-method}
\title{NCBI-to-Ensembl gene identifier mappings}
\usage{
NcbiToEnsembl(object, ...)

\S4method{NcbiToEnsembl}{integer}(object, organism, format = c("1:1", "long"))
}
\arguments{
\item{object}{Object.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{format}{\code{character(1)}.
Formatting method to apply:
\itemize{
\item \code{"1:1"}: \emph{Recommended.} Return with 1:1 mappings. For Ensembl genes that
don't map 1:1 with NCBI, pick the oldest NCBI identifier. Genes that don't
map to NCBI will contain \code{NA} in \code{ncbiGeneId} column.
\item \code{"long"}: Return \code{1:many} in long format.
}}

\item{...}{Additional arguments.}
}
\value{
\code{NcbiToEnsembl}.
}
\description{
NCBI-to-Ensembl gene identifier mappings
}
\note{
Updated 2023-09-16.
}
\examples{
## integer ====
x <- NcbiToEnsembl(object = c(1L, 2L), organism = "Homo sapiens")
print(x)
}
