% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcbiGeneHistory.R
\name{NcbiGeneHistory}
\alias{NcbiGeneHistory}
\title{NCBI gene history}
\usage{
NcbiGeneHistory(organism)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}
}
\value{
\code{NcbiGeneHistory}.
}
\description{
NCBI gene history
}
\note{
Updated 2023-09-26.
}
\examples{
## Homo sapiens.
object <- NcbiGeneHistory(organism = "Homo sapiens")
print(object)
}
