% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{NcbiGeneInfo-class}
\alias{NcbiGeneInfo-class}
\title{NCBI gene identifier information}
\value{
\code{NcbiGeneInfo}.
}
\description{
NCBI gene identifier information
}
\note{
Updated 2023-09-26.
}
