% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcbiGeneInfo.R
\name{NcbiGeneInfo}
\alias{NcbiGeneInfo}
\title{Import NCBI (Entrez) gene identifier information}
\usage{
NcbiGeneInfo(
  organism,
  taxonomicGroup = NULL,
  refseqGeneSummary = FALSE,
  goTerms = FALSE
)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{taxonomicGroup}{\code{character(1)}.
NCBI FTP server taxonomic group subdirectory path (e.g. "Mammalia").
Defining this manually avoids having to query the FTP server.}

\item{refseqGeneSummary}{\code{logical(1)}.
Include RefSeq gene summary in \code{"refseqGeneSummary"} column.
Requires Bioconductor GeneSummary package to be installed.}

\item{goTerms}{\code{logical(1)}.
Return nested gene ontology (GO) terms in \code{"goBp"} (BP: biological process),
\code{"goCc"} (CC: cellular component), and \code{"goMf"} (MF: molecular function)
columns. This is computationally intensive, and not supported for all
genomes. Intended primarily for \emph{Homo sapiens}.}
}
\value{
\code{NcbiGeneInfo}.
}
\description{
Import NCBI (Entrez) gene identifier information
}
\note{
Updated 2023-12-13.
}
\examples{
object <- NcbiGeneInfo(
    organism = "Homo sapiens",
    taxonomicGroup = "Mammalia"
)
print(object)
}
\seealso{
\itemize{
\item https://ftp.ncbi.nih.gov/gene/DATA/GENE_INFO/
}
}
