% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/NcbiToEnsembl-methods.R
\name{NcbiToEnsembl}
\alias{NcbiToEnsembl}
\alias{NcbiToEnsembl,Hgnc-method}
\alias{NcbiToEnsembl,integer-method}
\title{NCBI-to-Ensembl gene identifier mappings}
\usage{
NcbiToEnsembl(object, ...)

\S4method{NcbiToEnsembl}{Hgnc}(object)

\S4method{NcbiToEnsembl}{integer}(object, organism, strict = TRUE)
}
\arguments{
\item{object}{Object.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{strict}{\code{logical(1)}.
Error on any mismatches, otherwise return \code{NA}.}

\item{...}{Additional arguments.}
}
\value{
\code{NcbiToEnsembl}.
}
\description{
NCBI-to-Ensembl gene identifier mappings
}
\note{
Updated 2023-11-27.
}
\examples{
## integer ====
x <- NcbiToEnsembl(object = c(1L, 2L), organism = "Homo sapiens")
print(x)
}
