% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Shared parameter documentation}
\arguments{
\item{broadClass}{\code{logical(1)}.
Include broad class definitions.}

\item{seqinfo}{\code{Seqinfo} or \code{NULL}.}

\item{synonyms}{\code{logical(1)}.
Include gene synonyms.
Queries the Ensembl web server, and is CPU intensive.}
}
\value{
No value.
}
\description{
Shared parameter documentation
}
\keyword{internal}
