% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Protein2Gene-class}
\alias{Protein2Gene-class}
\title{Protein-to-gene mappings}
\value{
\code{Protein2Gene}.
}
\description{
Protein-to-gene mappings
}
\details{
Contains a \code{DataFrame} with \code{proteinId}, \code{geneId}, and \code{geneName} columns.
}
\note{
Updated 2021-02-02.
}
\section{Genome metadata}{


We recommend slotting \code{organism}, \code{genomeBuild}, and \code{ensemblRelease} into
\code{metadata()}.
}

