% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ProteinToGene-class}
\alias{ProteinToGene-class}
\title{Protein-to-gene mappings}
\value{
\code{ProteinToGene}.
}
\description{
Protein-to-gene mappings
}
\details{
Contains a \code{DFrame} with \code{proteinId}, \code{geneId}, and \code{geneName} columns.
}
\note{
Updated 2023-09-16.
}
\section{Genome metadata}{


We recommend slotting \code{organism}, \code{genomeBuild}, and \code{ensemblRelease} into
\code{metadata()}.
}

