% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{DataFrame}
\alias{Seqinfo}
\alias{genome}
\alias{pasteURL}
\alias{seqinfo}
\alias{seqlevels}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AcidBase}{\code{\link[AcidBase]{pasteURL}}}

  \item{AcidGenerics}{\code{\link[AcidGenerics:reexports]{DataFrame}}}

  \item{GenomeInfoDb}{\code{\link[GenomeInfoDb:Seqinfo-class]{Seqinfo}}, \code{\link[GenomeInfoDb:seqinfo]{genome}}, \code{\link[GenomeInfoDb]{seqinfo}}, \code{\link[GenomeInfoDb:seqinfo]{seqlevels}}}
}}

