% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RefseqExons-class}
\alias{RefseqExons-class}
\title{RefSeq exon annotations}
\value{
\code{RefseqExons}.
}
\description{
RefSeq exon annotations
}
\details{
Contains a \code{CompressedGRangesList} with RefSeq exon-level annotations.
}
\note{
Updated 2024-01-02.
}
