% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RefseqGenes-class}
\alias{RefseqGenes-class}
\title{RefSeq gene annotations}
\value{
\code{RefseqGenes}.
}
\description{
RefSeq gene annotations
}
\details{
Contains a \code{CompressedGRangesList} with RefSeq gene-level annotations.
}
\note{
Updated 2023-09-26.
}
