% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RefseqTranscripts-class}
\alias{RefseqTranscripts-class}
\title{RefSeq transcript annotations}
\value{
\code{RefseqTranscripts}.
}
\description{
RefSeq transcript annotations
}
\details{
Contains a \code{CompressedGRangesList} with RefSeq transcript-level annotations.
}
\note{
Updated 2023-09-26.
}
