% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/show-methods.R
\name{show}
\alias{show}
\alias{show,EnsemblGenes-method}
\alias{show,EnsemblTranscripts-method}
\alias{show,FlyBaseGenes-method}
\alias{show,FlyBaseTranscripts-method}
\alias{show,GencodeGenes-method}
\alias{show,GencodeTranscripts-method}
\alias{show,RefSeqGenes-method}
\alias{show,RefSeqTranscripts-method}
\alias{show,UCSCGenes-method}
\alias{show,UCSCTranscripts-method}
\alias{show,WormBaseGenes-method}
\alias{show,WormBaseTranscripts-method}
\title{Show an object}
\usage{
show(object)

\S4method{show}{EnsemblGenes}(object)

\S4method{show}{EnsemblTranscripts}(object)

\S4method{show}{FlyBaseGenes}(object)

\S4method{show}{FlyBaseTranscripts}(object)

\S4method{show}{GencodeGenes}(object)

\S4method{show}{GencodeTranscripts}(object)

\S4method{show}{RefSeqGenes}(object)

\S4method{show}{RefSeqTranscripts}(object)

\S4method{show}{UCSCGenes}(object)

\S4method{show}{UCSCTranscripts}(object)

\S4method{show}{WormBaseGenes}(object)

\S4method{show}{WormBaseTranscripts}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Console output.
}
\description{
Show an object
}
\note{
Updated 2023-04-26.
}
\author{
Michael Steinbaugh
}
