% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/stripExonVersions-methods.R
\name{stripExonVersions}
\alias{stripExonVersions}
\alias{stripExonVersions,Matrix-method}
\alias{stripExonVersions,character-method}
\alias{stripExonVersions,integer-method}
\alias{stripExonVersions,matrix-method}
\title{Strip exon identifier versions}
\usage{
stripExonVersions(object, ...)

\S4method{stripExonVersions}{Matrix}(object)

\S4method{stripExonVersions}{character}(object)

\S4method{stripExonVersions}{integer}(object)

\S4method{stripExonVersions}{matrix}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains gene identifiers without version numbers.
}
\description{
Strip exon identifier versions
}
\details{
Intentionally does not strip transcript versions, if defined in the object.
To perform that action in parallel, use \code{stripVersions} instead.
}
\note{
Updated 2024-01-02.

This method is strict, and will only strip Ensembl exon identifiers
beginning with \code{"ENS.*G"}.
}
\examples{
## Ensembl.
stripExonVersions("ENSG00000002586.20")

## GENCODE.
stripExonVersions("ENSG00000116288.13_10")
stripExonVersions("ENSG00000002586.20_PAR_Y")
}
\seealso{
\itemize{
\item https://www.ensembl.org/info/genome/stable_ids/index.html
}
}
