% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/stripGeneVersions-methods.R
\name{stripGeneVersions}
\alias{stripGeneVersions}
\alias{stripGeneVersions,Matrix-method}
\alias{stripGeneVersions,character-method}
\alias{stripGeneVersions,matrix-method}
\title{Strip gene identifier versions}
\usage{
stripGeneVersions(object, ...)

\S4method{stripGeneVersions}{Matrix}(object)

\S4method{stripGeneVersions}{character}(object)

\S4method{stripGeneVersions}{matrix}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains gene identifiers without version numbers.
}
\description{
Strip gene identifier versions
}
\details{
Intentionally does not strip transcript versions, if defined in the object.
To perform that action in parallel, use \code{stripVersions} instead.
}
\note{
Updated 2021-08-03.

This method is strict, and will only strip Ensembl gene identifiers
beginning with \code{"ENS.*G"}.
}
\examples{
## Ensembl.
stripGeneVersions("ENSG00000002586.20")

## GENCODE.
stripGeneVersions("ENSG00000002586.20_PAR_Y")
}
\seealso{
\itemize{
\item http://www.ensembl.org/info/genome/stable_ids/index.html
}
}
