% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/summary-methods.R
\name{summary}
\alias{summary}
\alias{summary,Gene2Symbol-method}
\alias{summary,Tx2Gene-method}
\title{Object summaries}
\usage{
summary(object, ...)

\S4method{summary}{Gene2Symbol}(object)

\S4method{summary}{Tx2Gene}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{NULL}.
}
\description{
Object summaries
}
\note{
Updated 2019-11-19.
}
\examples{
## Gene2Symbol ====
df <- DataFrame(
    "geneId" = c(
        "ENSG00000228572.7",
        "ENSG00000182378.14"
    ),
    "geneName" = c(
        "AL954722.1",
        "PLCXD1"
    )
)
g2s <- Gene2Symbol(df)
summary(g2s)

## Tx2Gene ====
df <- DataFrame(
    "txId" = c(
        "ENST00000635602.1",
        "ENST00000635506.1"
    ),
    "geneId" = c(
        "ENSG00000283061.1",
        "ENSG00000283061.1"
    )
)
t2g <- Tx2Gene(df)
summary(t2g)
}
\keyword{internal}
