% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TxToGene-class}
\alias{TxToGene-class}
\title{Transcript-to-gene mappings}
\value{
\code{TxToGene}.
}
\description{
Transcript-to-gene mappings
}
\details{
Contains a \code{DFrame} with \code{txId} and \code{geneId} columns.
}
\note{
Updated 2023-09-16.
}
\section{Genome metadata}{


We recommend slotting \code{organism}, \code{genomeBuild}, and \code{release} into
\code{metadata}.

Ensembl examples:
\itemize{
\item \code{organism}: "Homo sapiens".
\item \code{genomeBuild}: "GRCh38".
\item \code{release}: 100L.
}
}

