% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{UcscExons-class}
\alias{UcscExons-class}
\title{UCSC exon annotations}
\value{
\code{UcscExons}.
}
\description{
UCSC exon annotations
}
\details{
Contains \code{GRanges} with UCSC exon-level annotations.
}
\note{
Updated 2023-12-20.
}
