% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{UCSCGenes-class}
\alias{UCSCGenes-class}
\title{UCSC gene annotations}
\value{
\code{UCSCGenes}.
}
\description{
UCSC gene annotations
}
\details{
Contains \code{GenomicRanges} with UCSC gene-level annotations.
}
\note{
Updated 2021-10-14.
}
