% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{UcscTranscripts-class}
\alias{UcscTranscripts-class}
\title{UCSC transcript annotations}
\value{
\code{UcscTranscripts}.
}
\description{
UCSC transcript annotations
}
\details{
Contains \code{GRanges} with UCSC transcript-level annotations.
}
\note{
Updated 2023-09-26.
}
