% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateGeneSymbols.R
\name{updateGeneSymbols}
\alias{updateGeneSymbols}
\title{Update gene symbols}
\usage{
updateGeneSymbols(geneNames, organism)
}
\arguments{
\item{geneNames}{\code{character}.
Gene names (symbols).}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}
}
\value{
\code{character}.
Named character, with original input defined in the names.
}
\description{
Update gene symbols
}
\details{
Human genes are checked against HGNC.
Other organisms are checked against NCBI.

Intentionally errors on match failure.
}
\note{
Updated 2023-12-21.
}
\examples{
## Homo sapiens ====
object <- updateGeneSymbols(
    geneNames = c("ZCCHC6", "ZCCHC11"),
    organism = "Homo sapiens"
)
print(object)

## Mus musculus ====
object <- updateGeneSymbols(
    geneNames = c("Zcchc6", "Zcchc11"),
    organism = "Mus musculus"
)
print(object)
}
\seealso{
\itemize{
\item https://twitter.com/samuel_marsh/status/1737855289876201922
\item https://waldronlab.io/HGNChelper/
}
}
