% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/updateObject-methods.R
\name{updateObject}
\alias{updateObject}
\alias{updateObject,EnsemblGenes-method}
\title{Update object}
\usage{
updateObject(object, ..., verbose = FALSE)

\S4method{updateObject}{EnsemblGenes}(object, ..., verbose = FALSE)
}
\description{
Update object
}
\note{
Updated 2023-03-07.
}
\author{
Michael Steinbaugh
}
