% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{WormbaseExons-class}
\alias{WormbaseExons-class}
\title{WormBase exon annotations}
\value{
\code{WormbaseExons}.
}
\description{
WormBase exon annotations
}
\details{
Contains \code{GRanges} with WormBase exon-level annotations.
}
\note{
Updated 2023-12-20.
}
