% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{WormbaseGenes-class}
\alias{WormbaseGenes-class}
\title{WormBase gene annotations}
\value{
\code{WormbaseGenes}.
}
\description{
WormBase gene annotations
}
\details{
Contains \code{GRanges} with WormBase gene-level annotations.
}
\note{
Updated 2023-09-26.
}
