% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{WormbaseTranscripts-class}
\alias{WormbaseTranscripts-class}
\title{WormBase transcript annotations}
\value{
\code{WormbaseTranscripts}.
}
\description{
WormBase transcript annotations
}
\details{
Contains \code{GRanges} with WormBase transcript-level annotations.
}
\note{
Updated 2023-09-26.
}
