% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{AcidGenomesTestsUrl}
\alias{AcidGenomesTestsUrl}
\title{AcidGenomes test data URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
AcidGenomesTestsUrl
}
\description{
AcidGenomes test data URL
}
\note{
Updated 2023-10-04.
}
\examples{
AcidGenomesTestsUrl
}
\keyword{internal}
