% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{AcidGenomesTestsURL}
\alias{AcidGenomesTestsURL}
\title{AcidGenomes test data URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
AcidGenomesTestsURL
}
\description{
AcidGenomes test data URL
}
\note{
Updated 2020-10-06.
}
\examples{
AcidGenomesTestsURL
}
\keyword{internal}
