% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentGenomeBuild.R
\name{currentGenomeBuild}
\alias{currentGenomeBuild}
\alias{currentEnsemblBuild}
\alias{currentGencodeBuild}
\alias{currentRefSeqGenomeBuild}
\alias{currentUCSCGenomeBuild}
\title{Current genome build}
\usage{
currentEnsemblBuild(organism)

currentGencodeBuild(organism)

currentRefSeqGenomeBuild(organism, subdir = "vertebrate_mammalian")

currentUCSCGenomeBuild(organism)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{subdir}{\code{character(1)}.
\emph{Only applies to RefSeq}.
FTP server subdirectory path (e.g. "vertebrate_mammalian").}
}
\value{
\code{character(1)}.
Genome assembly build version.
}
\description{
Fetch the current genome build (assembly) version from online resources.
}
\note{
Updated 2020-01-05.
}
\examples{
currentEnsemblBuild("Homo sapiens")
currentGencodeBuild("Homo sapiens")
currentRefSeqGenomeBuild(
    organism = "Homo sapiens",
    subdir = "vertebrate_mammalian"
)
currentUCSCGenomeBuild("Homo sapiens")
}
\seealso{
\itemize{
\item \href{https://rest.ensembl.org}{Ensembl REST API}.
\item \href{https://genome.ucsc.edu/goldenPath/help/api.html}{UCSC REST API}.
\item \href{https://ftp.ncbi.nlm.nih.gov/genomes/refseq/}{RefSeq genomes}.
}
}
