% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentGenomeVersion.R
\name{currentGenomeVersion}
\alias{currentGenomeVersion}
\alias{currentEnsemblVersion}
\alias{currentGencodeVersion}
\alias{currentRefseqVersion}
\alias{currentFlybaseVersion}
\alias{currentWormbaseVersion}
\title{Current genome version}
\usage{
currentEnsemblVersion()

currentGencodeVersion(organism = c("Homo sapiens", "Mus musculus"))

currentRefseqVersion()

currentFlybaseVersion(dmel = FALSE)

currentWormbaseVersion()
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{dmel}{\code{logical(1)}.
Return \emph{Drosophila melanogaster} genome release version.}
}
\value{
\code{character(1)} or \code{integer(1)} when possible.
Release version.
}
\description{
Obtain the latest release version from various genome annotation sources.
}
\note{
Updated 2025-04-14.
}
\examples{
## Ensembl.
try({
    currentEnsemblVersion()
})

## GENCODE.
try({
    currentGencodeVersion(organism = "Homo sapiens")
})

## RefSeq.
try({
    currentRefseqVersion()
})

## WormBase.
try({
    currentWormbaseVersion()
})
}
\seealso{
Refer to the koopa package for Bash shell variants.
}
