% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadEnsemblGenome.R
\name{downloadEnsemblGenome}
\alias{downloadEnsemblGenome}
\title{Download Ensembl reference genome}
\usage{
downloadEnsemblGenome(
  organism,
  genomeBuild = NULL,
  release = NULL,
  outputDir = getwd(),
  cache = FALSE
)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
\emph{Note:} don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{release}{\code{integer(1)}.
Ensembl release version (e.g. \code{90}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}

\item{outputDir}{\code{character(1)}.
Output directory path.}

\item{cache}{\code{logical(1)}.
Cache URLs locally, using BiocFileCache internally.}
}
\value{
Invisible \code{list}.
}
\description{
Download Ensembl reference genome
}
\note{
Updated 2022-05-04.
}
\examples{
## This example is bandwidth intensive.
## > downloadEnsemblGenome(
## >     organism = "Homo sapiens",
## >     genomeBuild = "GRCh38",
## >     release = 100L
## > )
}
