% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRefSeqGenome.R
\name{downloadRefSeqGenome}
\alias{downloadRefSeqGenome}
\title{Download RefSeq reference genome}
\usage{
downloadRefSeqGenome(
  organism,
  taxonomicGroup = NULL,
  genomeBuild = NULL,
  type = c("all", "transcriptome", "genome", "none"),
  annotation = c("all", "gtf", "gff", "none"),
  outputDir = "."
)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{taxonomicGroup}{\code{character(1)}.
\emph{Only applies to RefSeq}.
FTP server taxonomic group subdirectory path (e.g. "vertebrate_mammalian").}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
Note: don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{type}{\code{character(1)}.
Genome type to download.}

\item{annotation}{\code{character(1)}.
Annotation format to download.}

\item{outputDir}{\code{character(1)}.
Output directory path.}
}
\description{
Download RefSeq reference genome
}
\note{
Updated 2021-01-08.
}
\examples{
## This example is bandwidth intensive.
## > downloadRefSeqGenome(
## >     organism = "Homo sapiens",
## >     taxonomicGroup = "vertebrate_mammalian"
## >     genomeBuild = "GRCh38",
## >     type = "transcriptome",
## >     annotation = "gtf"
## > )
}
\seealso{
\itemize{
\item \href{https://ftp.ncbi.nlm.nih.gov/genomes/refseq/}{RefSeq Genomes FTP server}
\item \href{https://www.ncbi.nlm.nih.gov/projects/genome/guide/human/}{Human Genome Resources at NCBI}
\item \href{https://www.ncbi.nlm.nih.gov/genome/doc/ftpfaq/}{Genomes Download (FTP) FAQ}
}
}
