% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadUCSCGenome.R
\name{downloadUCSCGenome}
\alias{downloadUCSCGenome}
\title{Download UCSC reference genome}
\usage{
downloadUCSCGenome(organism, genomeBuild = NULL, outputDir = ".")
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
Note: don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{outputDir}{\code{character(1)}.
Output directory path.}
}
\value{
Invisible \code{list}.
}
\description{
Download UCSC reference genome
}
\note{
Updated 2021-01-21.
}
\section{Genome}{

\itemize{
\item \verb{<GENOME_BUILD>.chrom.sizes}: Two-column tab-separated text file containing
assembly sequence names and sizes.
\item \verb{<GENOME_BUILD>.chromAlias.txt}: Sequence name alias file, one line for
each sequence name.  First column is sequence name followed by tab
separated alias names.
}
}

\section{Transcriptome}{

\itemize{
\item \code{mrna.fa.gz}: Human mRNA from GenBank. This sequence data is updated
regularly via automatic GenBank updates.
\item \code{refMrna.fa.gz}: RefSeq mRNA from the same species as the genome.
This sequence data is updated regularly via automatic GenBank updates.
}
}

\section{Gene annotations}{

:
This directory contains GTF files for the main gene transcript sets where
available. They are sourced from the following gene model tables:
ncbiRefSeq, refGene, ensGene, knownGene.
}

\examples{
## This example is bandwidth intensive.
## > downloadUCSCGenome(organism = "Homo sapiens")
}
\seealso{
\itemize{
\item \href{https://hgdownload.soe.ucsc.edu/goldenPath/hg38/bigZips/analysisSet/}{hg38 pinned analysis set (for NGS pipelines)}.
}
}
