% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{EnsemblGenes-class}
\alias{EnsemblGenes-class}
\title{Ensembl gene annotations}
\value{
\code{EnsemblGenes}.
}
\description{
Ensembl gene annotations
}
\details{
Contains a \code{GRanges} with Ensembl gene-level annotations.
}
\note{
Updated 2021-01-18.
}
