% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/EnsemblToNcbi-methods.R
\name{EnsemblToNcbi}
\alias{EnsemblToNcbi}
\alias{EnsemblToNcbi,EnsemblGenes-method}
\alias{EnsemblToNcbi,GencodeGenes-method}
\alias{EnsemblToNcbi,Hgnc-method}
\alias{EnsemblToNcbi,character-method}
\title{Ensembl-to-NCBI gene identifier mappings}
\usage{
EnsemblToNcbi(object, ...)

\S4method{EnsemblToNcbi}{EnsemblGenes}(object)

\S4method{EnsemblToNcbi}{GencodeGenes}(object)

\S4method{EnsemblToNcbi}{Hgnc}(object)

\S4method{EnsemblToNcbi}{character}(object, organism = NULL)
}
\arguments{
\item{object}{Object.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{...}{Additional arguments.}
}
\value{
\code{EnsemblToNcbi}.
}
\description{
Ensembl-to-NCBI gene identifier mappings
}
\note{
Updated 2023-11-27.
}
\examples{
## character ====
x <- EnsemblToNcbi(
    object = c("ENSG00000000005.6", "ENSG00000000003.16"),
    organism = "Homo sapiens"
)
print(x)
}
