% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/EnsemblToNcbi-methods.R
\name{EnsemblToNcbi}
\alias{EnsemblToNcbi}
\alias{EnsemblToNcbi,EnsemblGenes-method}
\alias{EnsemblToNcbi,GencodeGenes-method}
\alias{EnsemblToNcbi,character-method}
\title{Ensembl-to-NCBI gene identifier mappings}
\usage{
EnsemblToNcbi(object, ...)

\S4method{EnsemblToNcbi}{EnsemblGenes}(object, format = c("1:1", "long"))

\S4method{EnsemblToNcbi}{GencodeGenes}(object, format = c("1:1", "long"))

\S4method{EnsemblToNcbi}{character}(object, organism = NULL, format = c("1:1", "long"))
}
\arguments{
\item{object}{Object.}

\item{format}{\code{character(1)}.
Formatting method to apply:
\itemize{
\item \code{"1:1"}: \emph{Recommended.} Return with 1:1 mappings. For Ensembl genes that
don't map 1:1 with NCBI, pick the oldest NCBI identifier. Genes that don't
map to NCBI will contain \code{NA} in \code{ncbiGeneId} column.
\item \code{"long"}: Return \code{1:many} in long format.
}}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{...}{Additional arguments.}
}
\value{
\code{EnsemblToNcbi}.
}
\description{
Ensembl-to-NCBI gene identifier mappings
}
\note{
Updated 2023-10-04.
}
\examples{
## character ====
x <- EnsemblToNcbi(
    object = c("ENSG00000000003", "ENSG00000000005"),
    organism = "Homo sapiens"
)
print(x)
}
