% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{EnsemblTranscripts-class}
\alias{EnsemblTranscripts-class}
\title{Ensembl transcript annotations}
\value{
\code{EnsemblTranscripts}.
}
\description{
Ensembl transcript annotations
}
\details{
Contains \code{GRanges} with Ensembl transcript-level annotations.
}
\note{
Updated 2023-04-13.
}
