% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntrezGeneInfo.R
\name{EntrezGeneInfo}
\alias{EntrezGeneInfo}
\title{Import NCBI Entrez gene identifier information}
\usage{
EntrezGeneInfo(organism, taxonomicGroup = NULL)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{taxonomicGroup}{\code{character(1)}.
NCBI FTP server taxonomic group subdirectory path (e.g. "Mammalia").
Defining this manually avoids having to query the FTP server.}
}
\value{
\code{EntrezGeneInfo}.
}
\description{
Import NCBI Entrez gene identifier information
}
\note{
Updated 2021-02-13.
}
\examples{
object <- EntrezGeneInfo(
    organism = "Homo sapiens",
    taxonomicGroup = "Mammalia"
)
print(object)
}
\seealso{
\itemize{
\item ftp://ftp.ncbi.nih.gov/gene/DATA/GENE_INFO/
\item \code{\link[=geneSynonyms]{geneSynonyms()}}.
}
}
