% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-methods.R
\name{export}
\alias{export}
\alias{export,Tx2Gene,ANY,ANY-method}
\title{Objects exported from other packages}
\usage{
\S4method{export}{Tx2Gene,ANY,ANY}(object, con, format, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{character}.
File path(s).
}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{IRanges}{\code{\link[IRanges:AtomicList-class]{CharacterList}}, \code{\link[IRanges:DataFrameList-class]{DataFrameList}}, \code{\link[IRanges:AtomicList-class]{FactorList}}, \code{\link[IRanges:IRanges-constructor]{IRanges}}, \code{\link[IRanges:AtomicList-class]{IntegerList}}, \code{\link[IRanges:AtomicList-class]{LogicalList}}, \code{\link[IRanges:AtomicList-class]{NumericList}}, \code{\link[IRanges:AtomicList-class]{RleList}}, \code{\link[IRanges:DataFrameList-class]{SplitDataFrameList}}}

  \item{S4Vectors}{\code{\link[S4Vectors:DataFrame-class]{DataFrame}}, \code{\link[S4Vectors:Factor-class]{Factor}}, \code{\link[S4Vectors:LLint-class]{LLint}}, \code{\link[S4Vectors:List-class]{List}}, \code{\link[S4Vectors:Rle-class]{Rle}}, \code{\link[S4Vectors:SimpleList-class]{SimpleList}}}
}
}
\details{
The \code{Tx2Gene} method automatically disables writing of column names, which
is the intended input format for tximport.
}
\note{
Updated 2021-08-07.
}
\examples{
object <- DataFrame(
    "txId" = c(
        "tx0001",
        "tx0002",
        "tx0003",
        "tx0004"
    ),
    "geneId" = c(
        "gene0001",
        "gene0001",
        "gene0002",
        "gene0002"
    )
)
export(object, file = "tx2gene.csv")
unlink("tx2gene.csv")
}
