% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-methods.R
\name{export}
\alias{export}
\alias{export,Tx2Gene,ANY,ANY-method}
\title{Export}
\usage{
\S4method{export}{Tx2Gene,ANY,ANY}(object, con, format, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{character}.
File path(s).
}
\description{
Export
}
\details{
The \code{Tx2Gene} method automatically disables writing of column names, which
is the intended input format for tximport.
}
\note{
Updated 2021-10-14.
}
\examples{
object <- DataFrame(
    "txId" = c(
        "tx0001",
        "tx0002",
        "tx0003",
        "tx0004"
    ),
    "geneId" = c(
        "gene0001",
        "gene0001",
        "gene0002",
        "gene0002"
    )
)
export(object, file = "tx2gene.csv")
file.remove("tx2gene.csv")
}
