% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FlyBaseGenes-class}
\alias{FlyBaseGenes-class}
\title{FlyBase gene annotations}
\value{
\code{FlyBaseGenes}.
}
\description{
FlyBase gene annotations
}
\details{
Contains \code{GRanges} with FlyBase gene-level annotations.
}
\note{
Updated 2021-10-14.
}
