% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FlybaseTranscripts-class}
\alias{FlybaseTranscripts-class}
\title{FlyBase transcript annotations}
\value{
\code{FlybaseTranscripts}.
}
\description{
FlyBase transcript annotations
}
\details{
Contains \code{GRanges} with FlyBase transcript-level annotations.
}
\note{
Updated 2023-09-26.
}
