% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencodeReleaseHistory.R
\name{gencodeReleaseHistory}
\alias{gencodeReleaseHistory}
\title{GENCODE release history}
\usage{
gencodeReleaseHistory(organism)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}
}
\value{
\code{DataFrame}.
}
\description{
GENCODE release history
}
\note{
Updated 2023-02-10.
}
\examples{
object <- gencodeReleaseHistory(organism = "Homo sapiens")
}
