% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GencodeTranscripts-class}
\alias{GencodeTranscripts-class}
\title{GENCODE transcript annotations}
\value{
\code{GencodeTranscripts}.
}
\description{
GENCODE transcript annotations
}
\details{
Contains \code{GRanges} with GENCODE transcript-level annotations.
}
\note{
Updated 2021-10-14.
}
