% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSynonyms.R
\name{geneSynonyms}
\alias{geneSynonyms}
\title{Gene synonyms}
\usage{
geneSynonyms(
  organism = c("Homo sapiens", "Mus musculus", "Drosophila melanogaster"),
  idType = c("Entrez", "Ensembl", "HGNC", "OMIM")
)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}
}
\value{
\code{DataFrame}.
}
\description{
Look up gene synonyms from NCBI.
}
\note{
Updated 2021-02-10.
}
\section{\emph{Caenorhabditis elegans}}{


Synonym support for \emph{Caenorhabditis elegans} is poor on NCBI. Use the
\href{https://r.acidgenomics.com/packages/wormbase/}{WormBase} package instead.
}

\examples{
## > object <- geneSynonyms(organism = "Homo sapiens")
## > print(object)
}
