% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSynonyms.R
\name{geneSynonyms}
\alias{geneSynonyms}
\title{Gene synonyms}
\usage{
geneSynonyms(
  organism,
  taxonomicGroup = NULL,
  geneIdType = c("NCBI", "Ensembl", "HGNC", "OMIM", "Entrez")
)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{taxonomicGroup}{\code{character(1)}.
NCBI FTP server taxonomic group subdirectory path (e.g. "Mammalia").
Defining this manually avoids having to query the FTP server.}

\item{geneIdType}{\code{character(1)}.
Type of gene identifier to return in the \code{geneId} column.}
}
\value{
\code{DataFrame} containing \code{geneId} and \code{geneSynonyms} columns.
}
\description{
Look up gene synonyms from NCBI.
}
\note{
Updated 2023-03-01.
}
\section{\emph{Caenorhabditis elegans}}{


Synonym support for \emph{Caenorhabditis elegans} is poor on NCBI. Use the
\href{https://r.acidgenomics.com/packages/wormbase/}{WormBase} package instead.
}

\examples{
object <- geneSynonyms(
    organism = "Homo sapiens",
    taxonomicGroup = "Mammalia",
    geneIdType = "Ensembl"
)
print(object)
}
