% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGFFMetadata.R
\name{getGFFMetadata}
\alias{getGFFMetadata}
\title{Get metadata about a GFF file}
\usage{
getGFFMetadata(file)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}
}
\value{
\code{list}.
Containing values, when possible:
\itemize{
\item \code{directives}
\item \code{format}
\item \code{genomeBuild}
\item \code{gffVersion}
\item \code{md5}
\item \code{organism}
\item \code{provider}
\item \code{release}
\item \code{sha256}
}
}
\description{
Get metadata about a GFF file
}
\note{
Updated 2021-08-06.
}
\examples{
url <- pasteURL(
    "ftp.ensembl.org",
    "pub",
    "release-102",
    "gtf",
    "homo_sapiens",
    "Homo_sapiens.GRCh38.102.gtf.gz",
    protocol = "ftp"
)
x <- getGFFMetadata(url)
print(x)
}
\seealso{
\itemize{
\item \code{\link[=getGFFDirectives]{getGFFDirectives()}}.
}
}
