% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HGNC-class}
\alias{HGNC-class}
\title{Human Genome Organization (HUGO) Gene Nomenclature Committee (HGNC) metadata}
\value{
\code{HGNC}.
}
\description{
Human Genome Organization (HUGO) Gene Nomenclature Committee (HGNC) metadata
}
\note{
Updated 2023-09-26.
}
