% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGNC.R
\name{HGNC}
\alias{HGNC}
\title{Import HGNC complete set metadata}
\usage{
HGNC()
}
\value{
\code{HGNC}.
}
\description{
Import HGNC complete set metadata
}
\note{
Updated 2021-02-10.
}
\examples{
object <- HGNC()
print(object)
}
\seealso{
\itemize{
\item https://www.genenames.org/
\item https://www.genenames.org/download/statistics-and-files/
}
}
