% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanToMouse.R
\name{HumanToMouse}
\alias{HumanToMouse}
\title{Import human-to-mouse gene mappings}
\usage{
HumanToMouse(unique = TRUE)
}
\arguments{
\item{unique}{\code{logical(1)}.
Only return unique 1:1 mappings between human and mouse.
Recommended by default.}
}
\value{
\code{HumanToMouse}.
}
\description{
Import human-to-mouse gene mappings
}
\note{
Updated 2023-09-26.
}
\examples{
object <- HumanToMouse()
print(object)
}
\seealso{
\itemize{
\item https://www.informatics.jax.org/
\item https://www.biostars.org/p/9567892/
}
}
