% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEnsDbFromGFF.R
\name{makeEnsDbFromGFF}
\alias{makeEnsDbFromGFF}
\title{Make EnsDb object from a GFF/GTF file}
\usage{
makeEnsDbFromGFF(file)
}
\value{
\code{EnsDb}.
}
\description{
Make EnsDb object from a GFF/GTF file
}
\note{
Updated 2021-01-14.
}
\examples{
file <- pasteURL(AcidGenomesTestsURL, "ensembl.gtf")
edb <- makeEnsDbFromGFF(file = file)
}
