% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEnsDbFromGFF.R
\name{makeEnsDbFromGFF}
\alias{makeEnsDbFromGFF}
\alias{makeEnsDbFromGTF}
\title{Make EnsDb object from a GFF/GTF file}
\usage{
makeEnsDbFromGFF(file, organism = NULL, genomeBuild = NULL, release = NULL)

makeEnsDbFromGTF(file, organism = NULL, genomeBuild = NULL, release = NULL)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
Note: don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{release}{\code{integer(1)}.
Ensembl release version (e.g. \code{90}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}
}
\value{
\code{EnsDb}.
}
\description{
Wrapper for ensembldb importer functions.
}
\section{Functions}{
\itemize{
\item \code{makeEnsDbFromGFF}: Primary function.

\item \code{makeEnsDbFromGTF}: Alias for GTF files.
}}

\note{
Updated 2021-01-14.
}
\examples{
file <- pasteURL(AcidGenomesTestsURL, "ensembl.gtf")
edb <- makeEnsDbFromGFF(
    file = file,
    organism = "Homo sapiens",
    genomeBuild = "GRCh38",
    release = 100L
)
}
\seealso{
\itemize{
\item \code{ensembldb::ensDbFromGff()}.
\item \code{ensembldb::ensDbFromGtf()}.
}
}
