% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEnsDbFromGFF.R
\name{makeEnsDbFromGFF}
\alias{makeEnsDbFromGFF}
\title{Make EnsDb object from a GFF/GTF file}
\usage{
makeEnsDbFromGFF(file)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}
}
\value{
\code{EnsDb}.
}
\description{
Wrapper for ensembldb importer functions.
}
\note{
Updated 2023-04-26.
}
\examples{
## > file <- AcidBase::pasteURL(
## >     "ftp.ensembl.org",
## >     "pub",
## >     "release-102",
## >     "gtf",
## >     "homo_sapiens",
## >     "Homo_sapiens.GRCh38.102.gtf.gz",
## >     protocol = "https"
## > )
## > edb <- makeEnsDbFromGFF(file)
## > print(edb)
}
\seealso{
\itemize{
\item \code{ensembldb::ensDbFromGff()}.
\item \code{ensembldb::ensDbFromGtf()}.
}
}
