% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesFromEnsembl.R
\name{makeGRangesFromEnsembl}
\alias{makeGRangesFromEnsembl}
\alias{makeGRangesFromEnsDb}
\title{Make GenomicRanges from Ensembl}
\usage{
makeGRangesFromEnsembl(
  organism,
  level = c("genes", "transcripts"),
  genomeBuild = NULL,
  release = NULL,
  ignoreVersion = TRUE,
  synonyms = FALSE
)

makeGRangesFromEnsDb(
  object,
  level = c("genes", "transcripts"),
  ignoreVersion = TRUE,
  synonyms = FALSE
)
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{level}{\code{character(1)}.
Return as genes or transcripts.}

\item{genomeBuild}{\code{character(1)}.
Ensembl genome build assembly name (e.g. \code{"GRCh38"}). If set \code{NULL},
defaults to the most recent build available.
\emph{Note:} don't pass in UCSC build IDs (e.g. \code{"hg38"}).}

\item{release}{\code{integer(1)}.
Ensembl release version (e.g. \code{100}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}

\item{ignoreVersion}{\code{logical(1)}.
Ignore identifier (e.g. transcript, gene) versions.
When applicable, the identifier containing version numbers will be stored
in \code{txIdVersion} and \code{geneIdVersion}, and the variants without versions
will be stored in \code{txId}, \code{txIdNoVersion}, \code{geneId}, and \code{geneIdNoVersion}.}

\item{synonyms}{\code{logical(1)}.
Include gene synonyms.
Queries the Ensembl web server, and is CPU intensive.}

\item{object}{\code{EnsDb} or \code{character(1)}.
\code{EnsDb} object or name of specific annotation package containing a
versioned EnsDb object (e.g. "EnsDb.Hsapiens.v75").}
}
\value{
\code{EnsemblGenes} or \code{EnsemblTranscripts}.
}
\description{
Quickly obtain gene and transcript annotations from
\href{https://www.ensembl.org/}{Ensembl} using
\href{https://bioconductor.org/packages/AnnotationHub/}{AnnotationHub} and
\href{https://bioconductor.org/packages/ensembldb/}{ensembldb}.
}
\details{
Simply specify the desired organism, using the full Latin name. For example,
we can obtain human annotations with \verb{Homo sapiens}. Optionally, specific
Ensembl genome builds (e.g. \code{GRCh38}) and release versions (e.g. \code{87}) are
supported.

Under the hood, this function fetches annotations from AnnotationHub using
the ensembldb package. AnnotationHub supports versioned Ensembl releases,
back to version 87.

Genome build: use \code{"GRCh38"} instead of \code{"hg38"} for the genome build, since
we're querying Ensembl and not UCSC.
}
\section{Functions}{
\itemize{
\item \code{makeGRangesFromEnsembl}: Obtain annotations from Ensembl by
querying AnnotationHub.

\item \code{makeGRangesFromEnsDb}: Use a specific \code{EnsDb} object as the
annotation source. Alternatively, can pass in an EnsDb package name as
a \code{character(1)}.
}}

\note{
Updated 2022-05-04.
}
\section{Broad class definitions}{


For gene and transcript annotations, a \code{broadClass} column is added, which
generalizes the gene types into a smaller number of semantically-meaningful
groups:
\itemize{
\item \code{coding}.
\item \code{noncoding}.
\item \code{pseudo}.
\item \code{small}.
\item \code{decaying}.
\item \code{ig} (immunoglobulin).
\item \code{tcr} (T cell receptor).
\item \code{other}.
}
}

\section{GRCh37 (hg19) legacy annotations}{


\code{\link[=makeGRangesFromEnsembl]{makeGRangesFromEnsembl()}} supports the legacy \emph{Homo sapiens} GRCh37 (release
75) build by internally querying the \href{https://bioconductor.org/packages/EnsDb.Hsapiens.v75/}{EnsDb.Hsapiens.v75} package.
Alternatively, the corresponding GTF/GFF file can be loaded directly from
GENCODE or Ensembl.
}

\examples{
## Get annotations from Ensembl via AnnotationHub query.
genes <- makeGRangesFromEnsembl(
    organism = "Homo sapiens",
    level = "genes"
)
summary(genes)
transcripts <- makeGRangesFromEnsembl(
    organism = "Homo sapiens",
    level = "transcripts"
)
summary(transcripts)

## Get annotations from specific EnsDb object or package.
if (goalie::isInstalled("EnsDb.Hsapiens.v75")) {
    genes <- makeGRangesFromEnsDb(
        object = "EnsDb.Hsapiens.v75",
        level = "genes"
    )
    summary(genes)
}
}
\seealso{
\itemize{
\item \href{https://bioconductor.org/packages/AnnotationHub/}{AnnotationHub}.
\item \href{https://bioconductor.org/packages/ensembldb/}{ensembldb}.
\item \code{ensembldb::ensDbFromGff()}, \code{ensembldb::ensDbFromGtf()}.
}
}
