% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesFromGFF.R
\name{makeGRangesFromGFF}
\alias{makeGRangesFromGFF}
\title{Make GenomicRanges from a GFF/GTF file}
\usage{
makeGRangesFromGFF(
  file,
  level = c("genes", "transcripts"),
  ignoreVersion = TRUE,
  synonyms = FALSE
)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}

\item{level}{\code{character(1)}.
Return as genes or transcripts.}

\item{ignoreVersion}{\code{logical(1)}.
Ignore identifier (e.g. transcript, gene) versions.
When applicable, the identifier containing version numbers will be stored
in \code{txIdVersion} and \code{geneIdVersion}, and the variants without versions
will be stored in \code{txId}, \code{txIdNoVersion}, \code{geneId}, and \code{geneIdNoVersion}.}

\item{synonyms}{\code{logical(1)}.
Include gene synonyms.
Queries the Ensembl web server, and is CPU intensive.}
}
\value{
\code{GenomicRanges}.
}
\description{
Make GenomicRanges from a GFF/GTF file
}
\details{
Remote URLs and compressed files are supported.
}
\note{
Updated 2023-01-30.
}
\section{GFF/GTF specification}{


The GFF (General Feature Format) format consists of one line per feature,
each containing 9 columns of data, plus optional track definition lines.

The GTF (General Transfer Format) format is identical to GFF version 2.

The UCSC website has detailed conventions on the GFF3 format, including
the metadata columns.
}

\section{Feature type}{

\itemize{
\item \code{CDS}: \strong{C}o\strong{D}ing \strong{S}sequence. A contiguous sequence that contains a
genomic interval bounded by start and stop codons. CDS refers to the
portion of a genomic DNA sequence that is translated, from the start codon
to the stop codon.
\item \code{exon}: Genomic interval containing 5' UTR (\code{five_prime_UTR}), CDS, and
3' UTR (\code{three_prime_UTR}).
\item \code{mRNA}: Processed (spliced) mRNA transcript.
}

See also:
\itemize{
\item \href{https://daler.github.io/gffutils/}{gffutils documentation}
\item \href{https://www.ncbi.nlm.nih.gov/genbank/genomes_gff/}{GenBank GFF documentation}
\item \href{https://ccb.jhu.edu/software/stringtie/gff.shtml}{stringtie GFF documentation}
\item \href{http://mblab.wustl.edu/GTF2.html}{Brent Lab GTF2 spec notes}
\item \href{https://github.com/The-Sequence-Ontology/Specifications/blob/master/gff3.md}{Sequence Ontology GFF3 spec notes}
}
}

\section{Supported sources}{


Currently \code{\link[=makeGRangesFromGFF]{makeGRangesFromGFF()}} supports genomes from these sources:
\itemize{
\item Ensembl
\item GENCODE
\item RefSeq
\item UCSC
\item FlyBase
\item WormBase
}
}

\section{Ensembl}{


Note that \code{\link[=makeGRangesFromEnsembl]{makeGRangesFromEnsembl()}} offers native support for Ensembl genome
builds and returns additional useful metadata that isn't defined inside a
GFF/GTF file.

If you must load a GFF/GTF file directly, then use \code{\link[=makeGRangesFromGFF]{makeGRangesFromGFF()}}.

Example URLs:
\itemize{
\item Ensembl \emph{Homo sapiens} GRCh38.p13, release 108
\href{ftp://ftp.ensembl.org/pub/release-108/gtf/homo_sapiens/Homo_sapiens.GRCh38.108.gtf.gz}{GTF},
\href{ftp://ftp.ensembl.org/pub/release-108/gff3/homo_sapiens/Homo_sapiens.GRCh38.108.gff3.gz}{GFF3}
\item Ensembl \emph{Homo sapiens} GRCh37, release 108 (87)
\href{ftp://ftp.ensembl.org/pub/grch37/release-108/gtf/homo_sapiens/Homo_sapiens.GRCh37.87.gtf.gz}{GTF},
\href{ftp://ftp.ensembl.org/pub/grch37/release-108/gff3/homo_sapiens/Homo_sapiens.GRCh37.87.gff3.gz}{GFF3}
}
}

\section{GENCODE}{


Example URLs:
\itemize{
\item GENCODE \emph{Homo sapiens} GRCh38.p13, release 42
\href{ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_42/gencode.v42.annotation.gtf.gz}{GTF},
\href{ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_42/gencode.v42.annotation.gff3.gz}{GFF3}
\item GENCODE \emph{Homo sapiens} GRCh37, release 42
\href{ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_42/GRCh37_mapping/gencode.v42lift37.annotation.gtf.gz}{GTF},
\href{ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_42/GRCh37_mapping/gencode.v42lift37.annotation.gff3.gz}{GFF3}
\item GENCODE \emph{Mus musculus} GRCm39, release M31
\href{ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_mouse/release_M31/gencode.vM31.annotation.gtf.gz}{GTF},
\href{ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_mouse/release_M31/gencode.vM31.annotation.gff3.gz}{GFF3}
}
}

\section{GENCODE vs. Ensembl}{


Annotations available from Ensembl and GENCODE are very similar.

The GENCODE annotation is made by merging the manual gene annotation produced
by the Ensembl-Havana team and the Ensembl-genebuild automated gene
annotation. The GENCODE annotation is the default gene annotation displayed
in the Ensembl browser. The GENCODE releases coincide with the Ensembl
releases, although GENCODE can skip an Ensembl release if there is no update
to the annotation with respect to the previous release. In practical terms,
the GENCODE annotation is essentially identical to the Ensembl annotation.

However, GENCODE handles pseudoautosomal regions (PAR) differently than
Ensembl. The Ensembl GTF file only includes this annotation once, for
chromosome X. However, GENCODE GTF/GFF3 files include the annotation in the
PAR regions of both chromosomes. You'll see these genes contain a "_PAR_Y"
suffix.

Additionally, GENCODE GFF/GTF files import with a gene identifier containing
a suffix, which differs slightly from the Ensembl GFF/GTF spec
(e.g. GENCODE: \code{ENSG00000000003.14}; Ensembl: \code{ENSG00000000003}).

The \href{https://www.gencodegenes.org/pages/faq.html}{GENCODE FAQ} has additional
details.
}

\section{RefSeq}{


Refer to the
\href{ftp://ftp.ncbi.nlm.nih.gov/genomes/README_GFF3.txt}{current RefSeq spec}
for details.

Example URLs:
\itemize{
\item RefSeq \emph{Homo sapiens} GRCh38.p14
\href{ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/vertebrate_mammalian/Homo_sapiens/all_assembly_versions/GCF_000001405.40_GRCh38.p14/GCF_000001405.40_GRCh38.p14_genomic.gtf.gz}{GTF},
\href{ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/vertebrate_mammalian/Homo_sapiens/all_assembly_versions/GCF_000001405.40_GRCh38.p14/GCF_000001405.40_GRCh38.p14_genomic.gff.gz}{GFF3}
}

See also:
\itemize{
\item \href{https://www.ncbi.nlm.nih.gov/books/NBK50679/}{RefSeq FAQ}
\item ftp://ftp.ncbi.nih.gov/gene/DATA/gene2refseq.gz
}
}

\section{UCSC}{


Example URLs:
\itemize{
\item UCSC \emph{Homo sapiens} hg38 GTF files:
\href{https://hgdownload.cse.ucsc.edu/goldenPath/hg38/bigZips/genes/hg38.ensGene.gtf.gz}{hg38.ensGene.gtf.gz},
\href{https://hgdownload.cse.ucsc.edu/goldenPath/hg38/bigZips/genes/hg38.knownGene.gtf.gz}{hg38.knownGene.gtf.gz},
\href{https://hgdownload.cse.ucsc.edu/goldenPath/hg38/bigZips/genes/hg38.ncbiRefSeq.gtf.gz}{hg38.ncbiRefSeq.gtf.gz},
\href{https://hgdownload.cse.ucsc.edu/goldenPath/hg38/bigZips/genes/hg38.refGene.gtf.gz}{hg38.refGene.gtf.gz}
}

Related URLs:
\itemize{
\item \href{http://hgdownload.soe.ucsc.edu/downloads.html}{UCSC downloads}
\item \href{https://hgdownload.cse.ucsc.edu/goldenPath/hg38/bigZips/}{UCSC hg38 bigZips}
\item \href{https://genome.ucsc.edu/cgi-bin/hgTables}{UCSC hgTables}
}
}

\section{FlyBase}{


Example URLs:
\itemize{
\item FlyBase \emph{Drosophila melanogaster} r6.49
\href{ftp://ftp.flybase.net/releases/FB2022_06/dmel_r6.49/gtf/dmel-all-r6.49.gtf.gz}{GTF},
\href{ftp://ftp.flybase.net/releases/FB2022_06/dmel_r6.49/gff/dmel-all-r6.49.gff.gz}{GFF3}
}
}

\section{WormBase}{


Example URLs:
\itemize{
\item WormBase \emph{Caenorhabditis elegans} WS287
\href{ftp://ftp.wormbase.org/pub/wormbase/releases/WS287/species/c_elegans/PRJNA13758/c_elegans.PRJNA13758.WS287.canonical_geneset.gtf.gz}{GTF},
\href{ftp://ftp.wormbase.org/pub/wormbase/releases/WS287/species/c_elegans/PRJNA13758/c_elegans.PRJNA13758.WS287.annotations.gff3.gz}{GFF3}
}
}

\examples{
## Ensembl ====
file <- AcidBase::pasteURL(
    "ftp.ensembl.org",
    "pub",
    "release-108",
    "gtf",
    "homo_sapiens",
    "Homo_sapiens.GRCh38.108.gtf.gz",
    protocol = "ftp"
)
genes <- makeGRangesFromGFF(
    file = file,
    level = "genes",
    ignoreVersion = FALSE
)
summary(genes)
## > transcripts <- makeGRangesFromGFF(
## >     file = file,
## >     level = "transcripts",
## >     ignoreVersion = FALSE
## > )
## > summary(transcripts)

## GENCODE ====
## > file <- AcidBase::pasteURL(
## >     "ftp.ebi.ac.uk",
## >     "pub",
## >     "databases",
## >     "gencode",
## >     "Gencode_human",
## >     "release_42",
## >     "gencode.v42.annotation.gtf.gz",
## >     protocol = "ftp"
## > )
## > genes <- makeGRangesFromGFF(file = file, level = "genes")
## > summary(genes)
## > transcripts <- makeGRangesFromGFF(file = file, level = "transcripts")
## > summary(transcripts)

## RefSeq ====
## > file <- AcidBase::pasteURL(
## >     "ftp.ncbi.nlm.nih.gov",
## >     "genomes",
## >     "refseq",
## >     "vertebrate_mammalian",
## >     "Homo_sapiens",
## >     "all_assembly_versions",
## >     "GCF_000001405.40_GRCh38.p14",
## >     "GCF_000001405.40_GRCh38.p14_genomic.gff.gz",
## >     protocol = "ftp"
## > )
## > genes <- makeGRangesFromGFF(file = file, level = "genes")
## > summary(genes)
## > transcripts <- makeGRangesFromGFF(file = file, level = "transcripts")
## > summary(transcripts)

## UCSC ====
## > file <- AcidBase::pasteURL(
## >     "hgdownload.soe.ucsc.edu",
## >     "goldenPath",
## >     "hg38",
## >     "bigZips",
## >     "genes",
## >     "hg38.ensGene.gtf.gz",
## >     protocol = "ftp"
## > )
## > genes <- makeGRangesFromGFF(file = file, level = "genes")
## > summary(genes)
## > transcripts <- makeGRangesFromGFF(file = file, level = "transcripts")
## > summary(transcripts)
}
\seealso{
\itemize{
\item \code{rtracklayer::import()}.
\item \code{GenomicFeatures::makeTxDbFromGRanges()}.
\item \code{GenomicFeatures::makeTxDbFromGFF()}.
\item \code{GenomicFeatures:::.make_splicings()}.
\item \code{tximeta:::getRanges()}.
\item \code{AnnotationDbi::columns()}.
\item \code{GenomeInfoDb::GenomeDescription-class}, which describes useful \code{organism},
\code{commonName}, \code{providerVersion}, \code{provider}, and \code{releaseDate} accessors.
}
}
