% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesFromTxDb.R
\name{makeGRangesFromTxDb}
\alias{makeGRangesFromTxDb}
\title{Make GRanges from TxDb object}
\usage{
makeGRangesFromTxDb(
  object,
  level = c("transcripts", "genes", "exons", "cds"),
  ignoreVersion = TRUE
)
}
\value{
\code{GRanges}.
}
\description{
Make GRanges from TxDb object
}
\note{
Updated 2021-01-26.
}
\examples{
file <- pasteURL(
    "ftp.ensembl.org",
    "pub",
    "release-102",
    "gtf",
    "homo_sapiens",
    "Homo_sapiens.GRCh38.102.gtf.gz",
    protocol = "ftp"
)
txdb <- makeTxDbFromGFF(file)
gr <- makeGRangesFromTxDb(txdb)
print(gr)
}
