% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTxDbFromGFF.R
\name{makeTxDbFromGFF}
\alias{makeTxDbFromGFF}
\alias{makeTxDbFromGTF}
\title{Make TxDb from a GFF/GTF file}
\usage{
makeTxDbFromGFF(file, seqinfo, organism = NULL, source = NULL)

makeTxDbFromGTF(file, seqinfo, organism = NULL, source = NULL)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}

\item{seqinfo}{\code{Seqinfo}.
Information about the chromosomes.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{source}{\code{character(1)} or \code{NULL}.
Description of data source.
If left \code{NULL}, defaults to file name.}
}
\value{
\code{TxDb}.
}
\description{
Wrapper for GenomicFeatures \code{makeTxDbFromGFF} importer.
}
\details{
This step can be noisy and generate expected warnings, which are
intentionally suppressed:\preformatted{some exons are linked to transcripts not found in the file
The following orphan exon were dropped
The following orphan CDS were dropped
}
}
\section{Functions}{
\itemize{
\item \code{makeTxDbFromGFF}: Primary function.

\item \code{makeTxDbFromGTF}: Alias for GTF files.
}}

\note{
Updated 2021-01-14.

For Ensembl and GENCODE genomes, consider using
\code{makeEnsDbFromGFF} instead.
}
\examples{
## RefSeq.
gffFile <- pasteURL(
    "ftp.ncbi.nlm.nih.gov",
    "genomes",
    "refseq",
    "vertebrate_mammalian",
    "Homo_sapiens",
    "all_assembly_versions",
    "GCF_000001405.38_GRCh38.p12",
    "GCF_000001405.38_GRCh38.p12_genomic.gff.gz",
    protocol = "ftp"
)
reportFile <- sub(
    pattern = "_genomic\\\\.gff\\\\.gz$",
    replacement = "_assembly_report.txt",
    x = gffFile
)
seqinfo <- getRefSeqSeqinfo(reportFile)
txdb <- makeTxDbFromGFF(file = gffFile, seqinfo = seqinfo)
print(txdb)
}
\seealso{
\itemize{
\item \code{GenomicFeatures::makeTxDbFromGFF()}.
\item \code{GenomicFeatures::supportedMiRBaseBuildValues()}.
Note that \emph{Homo sapiens} GRCh38 isn't currently supported in mirbase.db.
\item \href{https://bioconductor.org/packages/TxDb.Hsapiens.UCSC.hg38.knownGene/}{TxDb.Hsapiens.UCSC.hg38.knownGene}.
}
}
