% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGencodeToEnsembl.R
\name{mapGencodeToEnsembl}
\alias{mapGencodeToEnsembl}
\title{Map GENCODE release to Ensembl}
\usage{
mapGencodeToEnsembl(release)
}
\arguments{
\item{release}{\code{integer(1)}.
Human GENCODE release (e.g. \code{39}).}
}
\value{
\code{integer(1)}.
Ensembl release.
}
\description{
Map GENCODE release to Ensembl
}
\note{
Updated 2022-01-17.
}
\examples{
mapGencodeToEnsembl(39L)
}
\seealso{
\itemize{
\item https://www.gencodegenes.org/
\item https://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/_README.TXT
}
}
