% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGeneNamesToEnsembl.R
\name{mapGeneNamesToEnsembl}
\alias{mapGeneNamesToEnsembl}
\title{Map gene names to Ensembl}
\usage{
mapGeneNamesToEnsembl(
  genes,
  organism,
  ignoreCase = FALSE,
  hgnc = NULL,
  ncbi = NULL
)
}
\arguments{
\item{genes}{Gene names (e.g. \code{"TUT4"}).}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{hgnc}{\code{Hgnc} object.
Supported for \emph{Homo sapiens} genome only. Snapshot of HGNC annotations.
When defined, overrides default behavior of mapping internally with
\code{mapGeneNamesToNcbi} to favor \code{mapGeneNamesToHgnc} instead.}

\item{ncbi}{\code{NcbiGeneInfo} object.
Snapshot of NCBI annotations. Passes to \code{mapGeneNamesToNcbi} internally.}
}
\description{
Map gene names to Ensembl
}
\details{
Internally matches using \code{mapGeneNamesToNcbi}, so we can support
gene synonym matching.
}
\note{
Updated 2025-04-14.
}
\examples{
## Homo sapiens.
x <- mapGeneNamesToEnsembl(
    genes = c("TUT4", "ZCCHC11", "TENT3A"),
    organism = "Homo sapiens"
)
print(x)

## Mus musculus
x <- mapGeneNamesToEnsembl(
    genes = c("Nfe2l2", "Nrf2"),
    organism = "Mus musculus"
)
print(x)
}
