% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGeneNamesToHgnc.R
\name{mapGeneNamesToHgnc}
\alias{mapGeneNamesToHgnc}
\title{Map gene names (symbols) to HGNC identifiers}
\usage{
mapGeneNamesToHgnc(genes, hgnc = NULL)
}
\arguments{
\item{genes}{\code{character}.
Human gene names (e.g. \code{"TUT4"}).}

\item{hgnc}{\code{Hgnc} or \code{NULL}.
If \code{NULL}, HGNC annotations will be downloaded automatically.}
}
\description{
Map gene names (symbols) to HGNC identifiers
}
\note{
Updated 2025-04-07.
}
\examples{
## Homo sapiens (only).
x <- mapGeneNamesToHgnc(genes = c("TUT4", "ZCCHC11", "TENT3A"))
print(x)
}
