% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGeneNamesToHGNC.R
\name{mapGeneNamesToHGNC}
\alias{mapGeneNamesToHGNC}
\title{Map gene names (symbols) to HGNC identifiers}
\usage{
mapGeneNamesToHGNC(genes, hgnc = NULL)
}
\arguments{
\item{genes}{\code{character}.
Human gene names (e.g. \code{"TUT4"}).}

\item{hgnc}{\code{HGNC} or \code{NULL}.
If \code{NULL}, HGNC annotations will be downloaded automatically.}
}
\description{
Map gene names (symbols) to HGNC identifiers
}
\note{
Updated 2023-03-03.
}
\examples{
## Homo sapiens (only).
x <- mapGeneNamesToHGNC(genes = c("TUT4", "ZCCHC11", "TENT3A"))
print(x)
}
