% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGoTerms.R
\name{mapGoTerms}
\alias{mapGoTerms}
\title{Map gene ontology (GO) identifiers to term names}
\usage{
mapGoTerms()
}
\value{
\code{DFrame}.
Contains \code{"id"} and \code{"name"} columns.
}
\description{
Map gene ontology (GO) identifiers to term names
}
\note{
Updated 2023-12-13.
}
\examples{
object <- mapGoTerms()
print(object)
}
\seealso{
\itemize{
\item Bioconductor GO.db package.
\item https://geneontology.org/docs/download-ontology/
\item https://www.biostars.org/p/9552810/
}
}
