% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mgi.R
\name{Mgi}
\alias{Mgi}
\title{Import Mouse Genome Informatics (MGI) metadata}
\usage{
Mgi()
}
\value{
\code{Mgi}.
}
\description{
Import Mouse Genome Informatics (MGI) metadata
}
\note{
Updated 2023-12-19.
}
\examples{
object <- Mgi()
print(object)
}
\seealso{
\itemize{
\item https://www.informatics.jax.org/
}
}
