% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGI.R
\name{MGI}
\alias{MGI}
\title{Import Mouse Genome Informatics (MGI) metadata}
\usage{
MGI()
}
\value{
\code{MGI}.
}
\description{
Import Mouse Genome Informatics (MGI) metadata
}
\note{
Updated 2023-03-01.
}
\examples{
object <- MGI()
print(object)
}
\seealso{
\itemize{
\item https://www.informatics.jax.org/
}
}
