% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Ncbi2Ensembl-methods.R
\name{Ncbi2Ensembl}
\alias{Ncbi2Ensembl}
\alias{Ncbi2Ensembl,integer-method}
\title{NCBI-to-Ensembl gene identifier mappings}
\usage{
Ncbi2Ensembl(object, ...)

\S4method{Ncbi2Ensembl}{integer}(object, organism, format = c("1:1", "long"))
}
\arguments{
\item{object}{Object.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{format}{\code{character(1)}.
Formatting method to apply:
\itemize{
\item \code{"1:1"}: \emph{Recommended.} Return with 1:1 mappings. For Ensembl genes that
don't map 1:1 with NCBI, pick the oldest NCBI identifier. Genes that don't
map to NCBI will contain \code{NA} in \code{ncbiGeneId} column.
\item \code{"long"}: Return \code{1:many} in long format.
}}

\item{...}{Additional arguments.}
}
\value{
\code{Ncbi2Ensembl}.
}
\description{
NCBI-to-Ensembl gene identifier mappings
}
\note{
Updated 2023-03-01.
}
\examples{
## integer ====
genes <- c(1L, 2L)
organism <- "Homo sapiens"
x <- Ncbi2Ensembl(object = genes, organism = organism)
print(x)
}
