% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcbiGeneHistory.R
\name{NcbiGeneHistory}
\alias{NcbiGeneHistory}
\title{NCBI gene history}
\usage{
NcbiGeneHistory(taxonomyId)
}
\arguments{
\item{`integer(1)`.}{NCBI taxonomy identifier.
For \emph{Homo sapiens}, use \code{9606L}.}
}
\value{
\code{DFrame}.
}
\description{
NCBI gene history
}
\note{
Updated 2023-09-14.
}
\examples{
## Homo sapiens.
df <- NcbiGeneHistory(taxonomyId = 9606L)
}
