% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/NcbiToEnsembl-methods.R
\name{NcbiToEnsembl}
\alias{NcbiToEnsembl}
\alias{NcbiToEnsembl,Hgnc-method}
\alias{NcbiToEnsembl,integer-method}
\title{NCBI-to-Ensembl gene identifier mappings}
\usage{
NcbiToEnsembl(object, ...)

\S4method{NcbiToEnsembl}{Hgnc}(object)

\S4method{NcbiToEnsembl}{integer}(object, organism, format = c("1:1", "long"), strict = TRUE)
}
\arguments{
\item{object}{Object.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{format}{\code{character(1)}.
Formatting method to apply:
\itemize{
\item \code{"1:1"}: \emph{Recommended.} Return with 1:1 mappings. For Ensembl genes that
don't map 1:1 with NCBI, pick the oldest NCBI identifier.
\item \code{"long"}: Return \code{1:many} in long format.
}}

\item{strict}{\code{logical(1)}.
Error on any mismatches, otherwise return \code{NA}.}

\item{...}{Additional arguments.}
}
\value{
\code{NcbiToEnsembl}.
}
\description{
NCBI-to-Ensembl gene identifier mappings
}
\note{
Updated 2023-09-16.
}
\examples{
## integer ====
x <- NcbiToEnsembl(object = c(1L, 2L), organism = "Homo sapiens")
print(x)
}
