% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/NcbiToEnsembl-methods.R
\name{NcbiToEnsembl}
\alias{NcbiToEnsembl}
\alias{NcbiToEnsembl,Hgnc-method}
\alias{NcbiToEnsembl,Mgi-method}
\alias{NcbiToEnsembl,integer-method}
\title{NCBI-to-Ensembl gene identifier mappings}
\usage{
NcbiToEnsembl(object, ...)

\S4method{NcbiToEnsembl}{Hgnc}(object)

\S4method{NcbiToEnsembl}{Mgi}(object)

\S4method{NcbiToEnsembl}{integer}(object, organism)
}
\arguments{
\item{object}{Object.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{...}{Additional arguments.}
}
\value{
\code{NcbiToEnsembl}.
}
\description{
NCBI-to-Ensembl gene identifier mappings
}
\note{
Updated 2023-11-27.
}
\examples{
## integer ====
x <- NcbiToEnsembl(object = c(2L, 1L), organism = "Homo sapiens")
print(x)
}
