% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/organism-methods.R
\name{organism}
\alias{organism}
\alias{organism<-}
\alias{organism,matrix-method}
\alias{organism,Matrix-method}
\alias{organism,data.frame-method}
\alias{organism,Annotated-method}
\alias{organism,DataFrame-method}
\alias{organism,GRanges-method}
\alias{organism<-,Annotated-method}
\title{Organism}
\usage{
organism(object)

organism(object) <- value

\S4method{organism}{matrix}(object)

\S4method{organism}{Matrix}(object)

\S4method{organism}{data.frame}(object)

\S4method{organism}{Annotated}(object)

\S4method{organism}{DataFrame}(object)

\S4method{organism}{GRanges}(object)

\S4method{organism}{Annotated}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}
}
\value{
\code{character(1)}.
Latin organism name (e.g. \emph{Homo sapiens}).
}
\description{
Organism
}
\note{
Updated 2021-02-01.
}
\examples{
## DataFrame ====
df <- S4Vectors::DataFrame(
    "txId" = c(
        "ENST00000635602.1",
        "ENST00000635506.1"
    ),
    "geneId" = c(
        "ENSG00000283061.1",
        "ENSG00000283061.1"
    ),
    row.names = c(
        "ENST00000635602.1",
        "ENST00000635506.1"
    )
)
x <- organism(df)
print(x)
}
\seealso{
[detectOrganism()][.
}
