% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RefSeqGenes-class}
\alias{RefSeqGenes-class}
\title{RefSeq gene annotations}
\value{
\code{RefSeqGenes}.
}
\description{
RefSeq gene annotations
}
\details{
Contains a \code{GRanges} with RefSeq gene-level annotations.
}
\note{
Updated 2021-01-10.
}
