% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RefSeqTranscripts-class}
\alias{RefSeqTranscripts-class}
\title{RefSeq transcript annotations}
\value{
\code{RefSeqTranscripts}.
}
\description{
RefSeq transcript annotations
}
\details{
Contains a \code{GRanges} with RefSeq transcript-level annotations.
}
\note{
Updated 2021-01-29.
}
