% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/show-methods.R
\name{show}
\alias{show}
\alias{show,EnsemblGenes-method}
\alias{show,EnsemblTranscripts-method}
\alias{show,FlybaseGenes-method}
\alias{show,FlybaseTranscripts-method}
\alias{show,GencodeGenes-method}
\alias{show,GencodeTranscripts-method}
\alias{show,RefseqGenes-method}
\alias{show,RefseqTranscripts-method}
\alias{show,UcscGenes-method}
\alias{show,UcscTranscripts-method}
\alias{show,WormbaseGenes-method}
\alias{show,WormbaseTranscripts-method}
\title{Show an object}
\usage{
show(object)

\S4method{show}{EnsemblGenes}(object)

\S4method{show}{EnsemblTranscripts}(object)

\S4method{show}{FlybaseGenes}(object)

\S4method{show}{FlybaseTranscripts}(object)

\S4method{show}{GencodeGenes}(object)

\S4method{show}{GencodeTranscripts}(object)

\S4method{show}{RefseqGenes}(object)

\S4method{show}{RefseqTranscripts}(object)

\S4method{show}{UcscGenes}(object)

\S4method{show}{UcscTranscripts}(object)

\S4method{show}{WormbaseGenes}(object)

\S4method{show}{WormbaseTranscripts}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Console output.
}
\description{
Show an object
}
\note{
Updated 2023-04-26.
}
\author{
Michael Steinbaugh
}
