% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/stripVersions-methods.R
\name{stripGeneVersions}
\alias{stripGeneVersions}
\alias{stripTranscriptVersions}
\alias{stripVersions}
\alias{stripGeneVersions,Matrix-method}
\alias{stripGeneVersions,character-method}
\alias{stripGeneVersions,matrix-method}
\alias{stripTranscriptVersions,Matrix-method}
\alias{stripTranscriptVersions,character-method}
\alias{stripTranscriptVersions,matrix-method}
\title{Strip gene or transcript identifier versions}
\usage{
stripGeneVersions(object, ...)

stripTranscriptVersions(object, ...)

\S4method{stripGeneVersions}{Matrix}(object)

\S4method{stripGeneVersions}{character}(object)

\S4method{stripGeneVersions}{matrix}(object)

\S4method{stripTranscriptVersions}{Matrix}(object)

\S4method{stripTranscriptVersions}{character}(object)

\S4method{stripTranscriptVersions}{matrix}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains transcript identifiers without version numbers.
}
\description{
Strip transcript identifier versions
}
\details{
Intentionally does not strip gene versions, if defined in the object.
To perform that action in parallel, use \code{stripVersions} instead.
}
\note{
Updated 2021-08-03.

This method is strict, and will only strip Ensembl transcript IDs
beginning with "ENS".
}
\examples{
## Ensembl.
stripGeneVersions("ENSG00000002586.20")
stripTranscriptVersions("ENST00000431238.7")

## GENCODE.
stripGeneVersions("ENSG00000002586.20_PAR_Y")
stripTranscriptVersions("ENST00000431238.7_PAR_Y")

## WormBase.
stripTranscriptVersions("cTel79B.1")
}
\seealso{
\itemize{
\item http://www.ensembl.org/info/genome/stable_ids/index.html
}
}
