% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/summary-methods.R
\name{summary}
\alias{summary}
\alias{summary,GeneToSymbol-method}
\alias{summary,TxToGene-method}
\title{Object summaries}
\usage{
summary(object, ...)

\S4method{summary}{GeneToSymbol}(object)

\S4method{summary}{TxToGene}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{NULL}.
}
\description{
Object summaries
}
\note{
Updated 2019-11-19.
}
\examples{
## GeneToSymbol ====
df <- S4Vectors::DataFrame(
    "geneId" = c(
        "ENSG00000228572.7",
        "ENSG00000182378.14"
    ),
    "geneName" = c(
        "AL954722.1",
        "PLCXD1"
    )
)
g2s <- GeneToSymbol(df)
summary(g2s)

## TxToGene ====
df <- S4Vectors::DataFrame(
    "txId" = c(
        "ENST00000635602.1",
        "ENST00000635506.1"
    ),
    "geneId" = c(
        "ENSG00000283061.1",
        "ENSG00000283061.1"
    )
)
t2g <- TxToGene(df)
summary(t2g)
}
\keyword{internal}
