% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Tx2Gene-class}
\alias{Tx2Gene-class}
\title{Transcript-to-gene identifier mappings}
\value{
\code{Tx2Gene}.
}
\description{
Transcript-to-gene identifier mappings
}
\details{
Contains a \code{DataFrame} with \code{txId} and \code{geneId} columns.
}
\note{
Updated 2021-08-10.
}
\section{Genome metadata}{


We recommend slotting \code{organism}, \code{genomeBuild}, and \code{release} into
\code{metadata}.

Ensembl examples:
\itemize{
\item \code{organism}: "Homo sapiens".
\item \code{genomeBuild}: "GRCh38".
\item \code{release}: 100L.
}
}

