% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Tx2Gene-methods.R
\name{Tx2Gene}
\alias{Tx2Gene}
\alias{Tx2Gene,DFrame-method}
\alias{Tx2Gene,GRanges-method}
\alias{Tx2Gene,GRangesList-method}
\alias{Tx2Gene,data.frame-method}
\alias{Tx2Gene,matrix-method}
\title{Transcript-to-gene identifier mappings}
\usage{
Tx2Gene(object, ...)

\S4method{Tx2Gene}{DFrame}(object, quiet = FALSE)

\S4method{Tx2Gene}{GRanges}(object, ...)

\S4method{Tx2Gene}{GRangesList}(object, ...)

\S4method{Tx2Gene}{data.frame}(object, ...)

\S4method{Tx2Gene}{matrix}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Arguments pass through to \code{DataFrame} method.}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}
}
\value{
\code{Tx2Gene}.
}
\description{
Transcript-to-gene identifier mappings
}
\note{
Updated 2021-08-10.
}
\examples{
## DataFrame ====
df <- DataFrame(
    "txId" = c(
        "ENST00000635602.1",
        "ENST00000635506.1"
    ),
    "geneId" = c(
        "ENSG00000283061.1",
        "ENSG00000283061.1"
    )
)
t2g <- Tx2Gene(df)
print(t2g)
}
\seealso{
\code{\link[=makeTx2Gene]{makeTx2Gene()}}.
}
