% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Tx2Gene-methods.R
\name{Tx2Gene}
\alias{Tx2Gene}
\alias{Tx2Gene,matrix-method}
\alias{Tx2Gene,data.frame-method}
\alias{Tx2Gene,DataFrame-method}
\alias{Tx2Gene,GRanges-method}
\alias{Tx2Gene,GRangesList-method}
\title{Transcript-to-gene identifier mappings}
\usage{
Tx2Gene(object, ...)

\S4method{Tx2Gene}{matrix}(object)

\S4method{Tx2Gene}{data.frame}(object)

\S4method{Tx2Gene}{DataFrame}(object)

\S4method{Tx2Gene}{GRanges}(object)

\S4method{Tx2Gene}{GRangesList}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Tx2Gene}.
}
\description{
Transcript-to-gene identifier mappings
}
\note{
No attempt is made to arrange the rows by transcript identifier.

Updated 2021-02-02.
}
\examples{
## DataFrame ====
df <- S4Vectors::DataFrame(
    "txId" = c(
        "ENST00000635602.1",
        "ENST00000635506.1"
    ),
    "geneId" = c(
        "ENSG00000283061.1",
        "ENSG00000283061.1"
    )
)
t2g <- Tx2Gene(df)
print(t2g)
}
\seealso{
\code{\link[=makeTx2Gene]{makeTx2Gene()}}.
}
