% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{UcscGenes-class}
\alias{UcscGenes-class}
\title{UCSC gene annotations}
\value{
\code{UcscGenes}.
}
\description{
UCSC gene annotations
}
\details{
Contains \code{GRanges} with UCSC gene-level annotations.
}
\note{
Updated 2023-09-26.
}
