% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{UCSCTranscripts-class}
\alias{UCSCTranscripts-class}
\title{UCSC transcript annotations}
\value{
\code{UCSCTranscripts}.
}
\description{
UCSC transcript annotations
}
\details{
Contains \code{GenomicRanges} with UCSC transcript-level annotations.
}
\note{
Updated 2021-10-14.
}
