% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/updateObject-methods.R
\name{updateObject}
\alias{updateObject}
\alias{updateObject,EnsemblGenes-method}
\alias{updateObject,EnsemblTranscripts-method}
\alias{updateObject,GencodeGenes-method}
\alias{updateObject,GencodeTranscripts-method}
\title{Update object}
\usage{
updateObject(object, ..., verbose = FALSE)

\S4method{updateObject}{EnsemblGenes}(object, ..., verbose = FALSE)

\S4method{updateObject}{EnsemblTranscripts}(object, ..., verbose = FALSE)

\S4method{updateObject}{GencodeGenes}(object, ..., verbose = FALSE)

\S4method{updateObject}{GencodeTranscripts}(object, ..., verbose = FALSE)
}
\description{
Update object
}
\note{
Updated 2023-03-07.
}
\author{
Michael Steinbaugh
}
