% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{WormBaseGenes-class}
\alias{WormBaseGenes-class}
\title{WormBase gene annotations}
\value{
\code{WormBaseGenes}.
}
\description{
WormBase gene annotations
}
\details{
Contains \code{GRanges} with WormBase gene-level annotations.
}
\note{
Updated 2021-10-14.
}
