% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{WormBaseTranscripts-class}
\alias{WormBaseTranscripts-class}
\title{WormBase transcript annotations}
\value{
\code{WormBaseTranscripts}.
}
\description{
WormBase transcript annotations
}
\details{
Contains \code{GRanges} with WormBase transcript-level annotations.
}
\note{
Updated 2021-10-14.
}
